- add & document `select` for `<select>` elements - update docs
