+++ docs/_modules/index.html
@@ -77,6 +77,7 @@
              
   <h1>All modules for which code is available</h1>
 <ul><li><a href="selenium_driverless/input/pointer.html">selenium_driverless.input.pointer</a></li>
+<li><a href="selenium_driverless/input/utils.html">selenium_driverless.input.utils</a></li>
 <li><a href="selenium_driverless/scripts/network_interceptor.html">selenium_driverless.scripts.network_interceptor</a></li>
 <li><a href="selenium_driverless/scripts/switch_to.html">selenium_driverless.scripts.switch_to</a></li>
 <li><a href="selenium_driverless/types/base_target.html">selenium_driverless.types.base_target</a></li>
