+++ docs/genindex.html
@@ -1017,6 +1017,8 @@
         <li><a href="api/Chrome.html#selenium_driverless.webdriver.Chrome.search_elements">(selenium_driverless.webdriver.Chrome method)</a>
 </li>
       </ul></li>
+      <li><a href="api/Input.html#selenium_driverless.input.utils.select">select() (in module selenium_driverless.input.utils)</a>
+</li>
       <li><a href="api/Context.html#selenium_driverless.types.context.Context.send_keys">send_keys() (selenium_driverless.types.context.Context method)</a>
 
       <ul>
