+++ docs/index.html
@@ -152,6 +152,7 @@ some methods aren’t documented yet properly</p>
 <li class="toctree-l2"><a class="reference internal" href="api/Input.html#selenium_driverless.input.pointer.MouseButton"><code class="docutils literal notranslate"><span class="pre">MouseButton</span></code></a></li>
 <li class="toctree-l2"><a class="reference internal" href="api/Input.html#selenium_driverless.input.pointer.Buttons"><code class="docutils literal notranslate"><span class="pre">Buttons</span></code></a></li>
 <li class="toctree-l2"><a class="reference internal" href="api/Input.html#selenium_driverless.input.pointer.EventType"><code class="docutils literal notranslate"><span class="pre">EventType</span></code></a></li>
+<li class="toctree-l2"><a class="reference internal" href="api/Input.html#select-element">Select Element</a></li>
 </ul>
 </li>
 <li class="toctree-l1"><a class="reference internal" href="api/RequestInterception.html">Request-Interception</a><ul>
