+++ docs/searchindex.js
@@ -1,1 +1,1 @@
-Search.setIndex({"docnames": ["api/By", "api/Chrome", "api/ChromeOptions", "api/Context", "api/Input", "api/RequestInterception", "api/Target", "api/WebELement", "index"], "filenames": ["api\\By.rst", "api\\Chrome.rst", "api\\ChromeOptions.rst", "api\\Context.rst", "api\\Input.rst", "api\\RequestInterception.rst", "api\\Target.rst", "api\\WebELement.rst", "index.rst"], "titles": ["By Element Locator", "webdriver.Chrome", "ChromeOptions", "Context", "Input", "Request-Interception", "Target", "WebElement", "Documentation of Driverless"], "terms": {"class": [0, 1, 2, 3, 4, 5, 6, 7], "selenium_driverless": [0, 1, 2, 3, 4, 5, 6, 7, 8], "type": [0, 1, 2, 3, 4, 5, 6, 7, 8], "sourc": [0, 1, 2, 3, 4, 5, 6, 7], "set": [0, 1, 2, 3, 5, 6, 7], "support": [0, 1, 2, 3, 6], "strategi": 0, "id": [0, 1, 5, 6, 7], "name": [0, 1, 2, 3, 5, 6, 7], "xpath": [0, 3, 6, 8], "tag_nam": [0, 3, 7], "tag": 0, "class_nam": [0, 7], "css_selector": 0, "css": 0, "selector": [0, 3, 6], "alia": [0, 1, 3, 6, 7], "option": [1, 2, 3, 4, 5, 6, 7, 8], "none": [1, 2, 3, 4, 5, 6, 7], "timeout": [1, 2, 3, 4, 5, 6, 7, 8], "30": [1, 3, 5, 6, 7], "debug": 1, "fals": [1, 2, 3, 5, 6, 7], "max_ws_siz": [1, 3, 5, 6], "134217728": 1, "control": 1, "chromium": [1, 2, 3, 6], "base": [1, 3, 6], "browser": [1, 2, 3, 5, 6], "without": [1, 3, 5], "ani": [1, 3, 5, 6, 7], "driver": [1, 2, 3, 4, 5, 6, 8], "properti": [1, 2, 3, 5, 6, 7], "frame_tre": 1, "dict": [1, 2, 3, 4, 5, 6, 7], "async": [1, 2, 3, 4, 5, 6, 7, 8], "all": [1, 2, 3, 6, 7, 8], "nest": [1, 7], "frame": [1, 3, 5, 6, 7], "within": [1, 3, 6], "current": [1, 3, 6, 7], "target": [1, 3, 4, 5, 7, 8], "str": [1, 2, 3, 4, 5, 6, 7], "targetinfo": [1, 3, 6, 8], "context": [1, 6, 8], "incognito": [1, 6], "new_context": 1, "proxy_bypass_list": 1, "proxy_serv": 1, "true": [1, 2, 3, 4, 5, 6, 7, 8], "universal_access_origin": 1, "url": [1, 2, 3, 5, 6, 7], "about": [1, 2], "blank": [1, 2], "creat": [1, 2, 3, 5], "new": [1, 3, 5], "paramet": [1, 2, 3, 4, 5, 6, 7], "first": [1, 2, 7], "tab": [1, 2, 3, 6], "start": [1, 2, 3, 5, 6], "default": [1, 2, 3, 4, 5, 6], "list": [1, 2, 3, 4, 5, 6, 7], "an": [1, 2, 3, 5, 6, 7], "origin": [1, 3, 5, 6], "grant": 1, "unlimit": 1, "cross": [1, 5], "access": [1, 6], "part": 1, "other": [1, 3, 5], "than": 1, "those": [1, 6], "constitut": 1, "ar": [1, 6], "ignor": [1, 2], "return": [1, 2, 3, 4, 5, 6, 7], "The": [1, 2, 3, 5, 6, 7], "proxi": [1, 2, 5], "doesn": [1, 2, 3, 5, 7], "t": [1, 2, 3, 5, 6, 7, 8], "work": 1, "window": [1, 3, 6, 7], "due": [1, 2, 4], "crbug": [1, 2], "1310057": 1, "server": [1, 2, 5], "us": [1, 2, 3, 5, 6, 7], "get_target": 1, "_type": 1, "context_id": [1, 3, 6, 7], "self": 1, "get": [1, 2, 3, 5, 6, 7, 8], "liter": [1, 3, 5, 6, 7], "page": [1, 3, 5, 6, 7, 8], "background_pag": 1, "service_work": 1, "param": [1, 5, 6], "filter": 1, "thi": [1, 2, 3, 5, 6, 7, 8], "function": [1, 3, 4, 6], "current_target": 1, "base_target": [1, 3, 6], "basetarget": [1, 5, 6, 8], "connect": [1, 2, 6, 7], "handl": [1, 3, 5, 7], "global": [1, 6], "onli": [1, 2, 3, 6, 7], "bind": 1, "cdp": [1, 3, 6], "protocol": [1, 3, 6], "mv3_extens": [1, 2], "background": [1, 3], "script": [1, 3, 6, 7, 8], "load": [1, 2, 3, 5, 6], "extens": [1, 2, 3, 5, 6], "manifest": 1, "version": 1, "3": [1, 3, 6, 7], "ensure_extensions_incognito_allow": 1, "ensur": [1, 7], "instal": 1, "allow": [1, 3, 6], "gener": [1, 7], "decid": 1, "whether": [1, 2, 3, 4, 5, 6, 7], "split": 1, "span": 1, "not_allow": 1, "configur": 1, "For": [1, 3, 6], "chang": [1, 5, 7], "behaviour": [1, 3], "you": [1, 2, 3, 5, 6, 7], "ll": 1, "have": [1, 6], "modifi": [1, 4, 8], "json": [1, 2, 3, 4, 5, 6, 7], "file": [1, 2, 3, 6, 7], "compress": [1, 2, 5], "directori": [1, 2, 3, 6], "see": [1, 3, 5, 6, 7, 8], "develop": [1, 6], "com": [1, 2, 8], "doc": [1, 6], "refer": [1, 3, 5, 6, 7], "base_context": 1, "which": [1, 2, 4, 5, 6], "isn": [1, 7], "downloads_dir": [1, 2, 3, 6], "download": [1, 2, 3, 6], "set_download_behaviour": [1, 3], "path": [1, 2, 3, 4, 6, 7], "deni": [1, 3], "allowandnam": [1, 3], "instead": [1, 2, 3, 6, 7], "can": [1, 2, 3, 6, 7], "caus": [1, 3, 5], "some": [1, 3, 5, 8], "bug": [1, 3], "current_context": 1, "switch": [1, 3], "target_id": [1, 3, 6], "2": [1, 3, 4, 5, 6, 7, 8], "targetid": [1, 6], "advanc": 1, "usag": [1, 3, 6], "float": [1, 3, 4, 6, 7], "second": [1, 3, 6, 7], "": [1, 7], "track": 1, "alreadi": [1, 5], "get_target_for_ifram": [1, 6], "ifram": [1, 3, 6, 7], "specif": [1, 3, 4, 6], "oopif": 1, "site": [1, 5], "isol": [1, 6], "solut": 1, "look": 1, "webel": [1, 3, 4, 6, 8], "content_docu": [1, 6, 7], "get_targets_for_ifram": [1, 6], "more": 1, "inform": 1, "wait_download": [1, 3, 6], "wait": [1, 3, 6, 7, 8], "someth": [1, 3, 6], "like": [1, 3, 6, 7], "frameid": [1, 3, 5, 6], "2d543b5e8b14945b280c537a4882a695": [1, 3, 6], "guid": [1, 3, 6], "c91df4d5": [1, 3, 6], "9b45": [1, 3, 6], "4962": [1, 3, 6], "84df": [1, 3, 6], "3749bd3f926d": [1, 3, 6], "http": [1, 2, 3, 5, 6, 7, 8], "www": [1, 2, 3, 6], "w3": [1, 2, 3, 5, 6], "org": [1, 2, 3, 5, 6, 7], "wai": [1, 2, 3, 6], "er": [1, 2, 3, 6], "test": [1, 2, 3, 6, 7], "xhtml": [1, 2, 3, 6], "testfil": [1, 2, 3, 6], "resourc": [1, 2, 3, 5, 6], "pdf": [1, 2, 3, 6], "dummi": [1, 2, 3, 6], "suggestedfilenam": [1, 3, 6], "specifi": [1, 3, 4, 5, 6], "guid_fil": [1, 2, 3, 6], "d": [1, 3, 6], "system": [1, 3, 6], "appdata": [1, 3, 6], "pycharm": [1, 3, 6], "scratch": [1, 3, 6], "time": [1, 3, 4, 5, 6, 7], "from": [1, 2, 3, 5, 6, 7, 8], "yet": [1, 2, 3, 6, 7, 8], "referr": [1, 3, 5, 6], "wait_load": [1, 3, 5, 6, 8], "web": [1, 3, 6, 7], "bool": [1, 2, 3, 4, 5, 6, 7], "webpag": [1, 3, 6, 7], "maximum": [1, 3, 6, 7], "same": [1, 3, 5, 6], "initi": [1, 3, 5, 6], "titl": [1, 3, 6, 8], "current_point": [1, 3], "pointer": [1, 3, 4, 6, 8], "send_kei": [1, 3, 6, 7], "text": [1, 3, 5, 6, 7, 8], "send": [1, 3, 6, 7], "kei": [1, 3, 5, 6, 7], "execute_raw_script": [1, 3, 6, 7], "arg": [1, 3, 6, 7], "await_r": [1, 3, 6], "serial": [1, 3, 6, 7], "deep": [1, 6], "max_depth": [1, 3, 6, 7], "execution_context_id": [1, 3, 6, 7], "unique_context": [1, 3, 6, 7], "execut": [1, 2, 3, 6, 7], "javascript": [1, 3, 6, 7], "globalthi": [1, 3, 6, 7], "argument": [1, 2, 3, 6, 7], "document": [1, 3, 6, 7], "obj": [1, 3, 6, 7], "here": [1, 3, 6, 7], "string": [1, 5, 6], "pass": [1, 2, 5, 6, 7], "either": [1, 2, 4, 6], "serializ": [1, 6], "remoteobject": [1, 6], "await": [1, 2, 3, 4, 5, 6, 7, 8], "valu": [1, 2, 3, 5, 6, 7], "idonli": [1, 6], "one": [1, 3, 5, 6, 7], "int": [1, 2, 3, 4, 5, 6, 7], "depth": [1, 6, 7], "object": [1, 3, 5, 6], "complet": [1, 6, 8], "run": [1, 2, 5, 6, 8], "exclus": [1, 6], "runtim": [1, 6, 7], "callfunctionon": [1, 6], "execute_script": [1, 3, 6, 7], "synchron": [1, 3, 6, 7], "descript": [1, 3, 6, 7], "execute_async_script": [1, 3, 6, 7], "asynchron": [1, 3, 6, 7], "i": [1, 2, 3, 5, 6, 7, 8], "recommend": [1, 3, 4, 7], "except": [1, 3, 7], "correctli": [1, 3, 7], "eval_async": [1, 3, 6, 7], "resolv": [1, 3, 6, 7], "length": [1, 3, 6, 7], "1": [1, 3, 4, 5, 6, 7], "re": [1, 3, 6, 7], "fetch": [1, 3, 5, 6, 7], "httpbin": [1, 3, 5, 6, 7], "mind": [1, 3, 5, 6, 7], "cor": [1, 3, 6, 7], "current_url": [1, 3, 6], "page_sourc": [1, 3, 6], "html": [1, 6, 7, 8], "close": [1, 3, 6, 7], "focu": [1, 3, 6, 7], "focus": [1, 7], "quit": [1, 3], "clean_dir": 1, "clean": [1, 2], "out": 1, "user": [1, 2, 5, 6, 7], "data": [1, 2, 3, 5, 6, 7], "dir": [1, 2], "current_target_info": 1, "current_window_handl": [1, 3], "deprec": [1, 2], "remov": [1, 2, 3, 6, 7], "current_window_id": 1, "windowid": 1, "belong": 1, "window_handl": [1, 3], "aren": [1, 6, 8], "order": [1, 5], "posit": [1, 3, 4], "do": [1, 7], "reli": 1, "index": 1, "iter": [1, 3, 5, 6], "them": 1, "new_window": [1, 3], "type_hint": [1, 3], "activ": [1, 3, 6], "should": [1, 2, 3, 4, 5, 6], "explicitli": 1, "new_target": 1, "set_window_st": 1, "state": 1, "normal": [1, 4], "minim": [1, 3], "maxim": [1, 3, 6], "fullscreen": 1, "normalize_window": 1, "size": [1, 7], "maximize_window": [1, 3], "fullscreen_window": [1, 3], "enter": [1, 5], "minimize_window": [1, 3], "rtype": [1, 2], "throttl": 1, "print_pag": [1, 3, 6], "print": [1, 2, 5, 6, 8], "switch_to": [1, 3, 8], "switchto": [1, 3, 8], "back": [1, 3, 4, 6], "goe": [1, 3, 6], "step": [1, 3, 6], "backward": [1, 3, 6], "histori": [1, 3, 6], "ha": [1, 2, 5, 6, 7], "forward": [1, 3, 4, 6], "refresh": [1, 3, 6], "get_cooki": [1, 3, 6], "cooki": [1, 3, 6], "singl": [1, 2, 3, 5, 6], "found": [1, 3, 6], "delete_cooki": [1, 3, 6], "domain": [1, 3, 6], "delet": [1, 3, 6], "given": [1, 3, 6, 7], "delete_all_cooki": [1, 3, 6], "add_cooki": [1, 3, 6], "cookie_dict": [1, 3, 6], "add": [1, 2, 3, 6], "network": [1, 3, 5, 6], "cookieparam": [1, 6], "static": [1, 5], "sleep": [1, 4, 8], "time_to_wait": 1, "sync": 1, "find_el": [1, 3, 6, 7, 8], "find": [1, 3, 6, 7], "element": [1, 3, 6, 7, 8], "locat": [1, 3, 4, 6, 7, 8], "By": [1, 2, 3, 5, 6, 7, 8], "actual": [1, 3, 6, 7], "queri": [1, 3, 6, 7], "how": [1, 3, 4, 6, 7], "long": [1, 3, 5, 6, 7], "exist": [1, 3, 6, 7, 8], "multipl": [1, 2, 3, 6, 7], "now": 6, "being": [1, 2, 3, 6, 7], "reload": [1, 3, 6], "loop": [1, 3, 6, 7], "search_el": [1, 3, 6], "similarli": 1, "ctrl": [1, 4], "f": 1, "devtool": [1, 2, 3, 6], "consol": 1, "plain": [1, 3, 6], "get_screenshot_as_fil": [1, 3, 6], "filenam": [1, 3, 6, 7], "save": [1, 2, 3, 6, 7], "screenshot": [1, 3, 6, 7], "png": [1, 3, 6, 7], "imag": [1, 3, 6, 7], "wish": [1, 3, 6], "your": [1, 3, 5, 6], "end": [1, 3, 6], "save_screenshot": [1, 3, 6], "func": [1, 3, 6], "get_screenshot_as_png": [1, 3, 6], "binari": [1, 2, 3, 5, 6, 7], "byte": [1, 3, 5, 6, 7], "snapshot": [1, 3, 6], "mhtml": [1, 3, 6], "save_snapshot": [1, 3, 6], "full": [1, 3, 6], "get_snapshot": [1, 3, 6], "set_window_s": [1, 3], "width": [1, 3], "height": [1, 3], "unless": [1, 2], "windowhandl": [1, 3], "pixel": [1, 3], "get_window_s": [1, 3], "code": [1, 5], "block": [1, 5], "1280": 1, "720": 1, "set_window_posit": [1, 3], "x": [1, 3, 4], "y": [1, 3, 4], "coordin": [1, 3, 4, 7], "get_window_posit": [1, 3], "0": [1, 3, 4, 5, 7, 8], "get_window_rect": [1, 3], "windowst": 1, "set_window_rect": [1, 3], "get_network_condit": [1, 3, 6], "emul": [1, 3, 6], "latenc": [1, 3, 6], "4": [1, 3, 4, 6], "download_throughput": [1, 3, 6], "upload_throughput": [1, 3, 6], "offlin": [1, 3, 6], "set_network_condit": [1, 3, 6], "connection_typ": [1, 3, 6], "addit": [1, 3, 6], "m": [1, 3, 6, 8], "throughput": [1, 3, 6], "500": [1, 3, 6], "1024": [1, 3, 6], "exampl": [1, 2, 3, 6, 8], "cellular2g": 1, "cellular3g": 1, "cellular4g": 1, "bluetooth": 1, "ethernet": 1, "wifi": [1, 3, 6], "wimax": 1, "both": [1, 3, 6], "upload": [1, 3, 6], "delete_network_condit": [1, 3, 6], "reset": [1, 3, 6], "set_permiss": [1, 3], "applic": [1, 3, 5], "permiss": [1, 3], "item": [1, 3], "prompt": 1, "appli": [1, 2], "clipboard": [1, 3], "read": [1, 3], "set_proxi": 1, "proxy_config": 1, "dynam": [1, 6], "mode": [1, 6], "fixed_serv": 1, "rule": 1, "proxyforhttp": 1, "scheme": [1, 5], "host": [1, 2, 6], "port": [1, 2], "proxyforftp": 1, "fallbackproxi": 1, "bypasslist": 1, "local": 1, "authentif": [1, 5], "set_auth": 1, "set_single_proxi": 1, "bypass_list": 1, "user1": [1, 2], "passwrd1": [1, 2], "5001": [1, 2], "when": [1, 2, 3, 5, 6], "been": [1, 2, 5], "driverless": [1, 2], "socks5": [1, 2], "authent": [1, 2, 5, 6], "1309413": [1, 2], "clear_proxi": 1, "clear": [1, 7], "usernam": [1, 5], "password": [1, 5, 6], "host_with_port": 1, "format": [1, 2, 7], "clear_auth": 1, "auth": [1, 2, 5], "wait_for_cdp": [1, 3, 6, 8], "event": [1, 3, 4, 5, 6, 7], "add_cdp_listen": [1, 3, 6], "callback": [1, 3, 5, 6], "listen": [1, 3, 6, 7], "remove_cdp_listen": [1, 3, 6], "get_cdp_event_it": [1, 3, 6], "over": [1, 3, 5, 6, 7], "asynciter": [1, 3, 5, 6], "execute_cdp_cmd": [1, 3, 6], "cmd": [1, 3, 6], "cmd_arg": [1, 3, 6], "10": [1, 3, 6, 8], "command": [1, 3, 4, 6, 7], "messag": [1, 3, 6], "Not": [1, 2, 3, 7], "receiv": [1, 3, 6], "kwarg": [1, 3, 4], "j": [1, 3, 6], "request": [1, 2, 3, 6, 8], "xhr": [1, 3, 6], "xmlhttprequest": [1, 3, 6], "get_sink": [1, 3, 6], "A": [1, 3, 6, 7], "sink": [1, 3, 6], "avail": [1, 3, 6, 7], "cast": [1, 3, 6], "get_issue_messag": [1, 3, 6], "error": [1, 3, 5, 6], "issu": [1, 3, 6], "session": [1, 2, 3, 6], "set_sink_to_us": [1, 3, 6], "sink_nam": [1, 3, 6], "its": [1, 3, 6], "start_desktop_mirror": [1, 3, 6], "desktop": [1, 3, 6], "mirror": [1, 3, 6], "start_tab_mirror": [1, 3, 6], "stop_cast": [1, 3, 6], "stop": [1, 3, 6], "alert": [1, 3, 8], "get_alert": 1, "5": [1, 2, 3, 4, 6, 7, 8], "default_cont": [1, 3], "frame_refer": 1, "union": [1, 2, 4, 5, 7], "bring": [1, 3, 6], "front": [1, 3, 6], "window_id": 1, "what": [1, 3], "kind": [1, 3], "webdriv": [2, 3, 5, 6, 8], "reus": 2, "use_extens": 2, "don": 2, "chrome": [2, 3, 5, 6, 7, 8], "while": 2, "requir": [2, 3, 4, 5], "ad": 2, "As": 2, "altern": 2, "might": [2, 5, 6, 7], "intercept": [2, 8], "provid": [2, 3, 5, 6], "add_argu": 2, "launch": 2, "pref": 2, "prefer": 2, "update_pref": 2, "updat": 2, "dot": 2, "user_data_dir": 2, "temporarili": 2, "temp": 2, "_dir": 2, "o": 2, "getcwd": 2, "isdir": 2, "mkdir": 2, "plugin": 2, "always_open_pdf_extern": 2, "download_data": 2, "absolut": [2, 7], "headless": 2, "startup_url": 2, "binary_loc": 2, "env": 2, "subprocess": 2, "popen": 2, "environ": 2, "add_extens": 2, "zip": 2, "crx": 2, "etc": [2, 6], "extract": 2, "debugger_address": 2, "address": 2, "remot": 2, "instanc": [2, 6, 7], "make": [2, 3, 5, 6], "well": [2, 3], "single_proxi": 2, "auto_clean_dir": 2, "automat": [2, 5, 7], "enable_mobil": 2, "android_packag": 2, "android": 2, "android_act": 2, "device_seri": 2, "enabl": [2, 7], "mobil": 2, "packag": 2, "implement": [2, 7], "accept_insecure_cert": 2, "accept": [2, 6, 7, 8], "insecur": 2, "certif": 2, "notimpl": [2, 7], "ignore_local_proxy_environment_vari": 2, "call": [2, 7], "http_proxi": 2, "https_proxi": 2, "pick": 2, "up": [2, 4], "add_experimental_opt": 2, "experiment": 2, "method": [2, 3, 5, 6, 7, 8], "is_incognito": 3, "1048576": [3, 6], "drive": 3, "chromedriv": 3, "behavior": [3, 5], "click": [3, 4, 7, 8], "const": 3, "accord": 3, "obj_id": [3, 7], "docs_sourc": 3, "start_monoton": 3, "everi": [3, 7], "associ": 3, "invok": 3, "manag": 3, "screen": 3, "oper": [3, 7], "take": [3, 4, 5, 6, 7], "best": [3, 6], "effort": [3, 6], "contain": 3, "active_el": 3, "frame_nam": 3, "parent_fram": 3, "main": [3, 4, 5, 8], "dictionari": [3, 6, 7], "correspond": [3, 5, 6], "visibl": [3, 6, 7], "my_cooki": 3, "scope": [3, 6], "secur": 3, "httponli": 3, "expiri": 3, "samesit": 3, "foo": 3, "bar": 3, "strict": [3, 5], "search": [3, 6], "resizeto": 3, "800": 3, "600": 3, "moveto": 3, "w3c": 3, "compat": 3, "100": 3, "200": [3, 5], "network_condit": [3, 6], "condit": [3, 6], "note": [3, 5, 6], "pointer_typ": 4, "mous": [4, 7], "down": 4, "press": 4, "pointerev": [4, 8], "releas": 4, "x_or_elem": 4, "move_to": [4, 7, 8], "move_kwarg": 4, "click_kwarg": [4, 7], "move": [4, 7], "basepoint": [4, 8], "total_tim": 4, "accel": 4, "mid_tim": 4, "smooth_soft": 4, "20": [4, 6], "pinter": 4, "total": 4, "acceler": 4, "deceler": 4, "pointermov": 4, "perform": [4, 7], "where": [4, 5], "half": 4, "curvi": 4, "line": 4, "move_path": 4, "dispatch": 4, "between": 4, "mouseup": [4, 7], "mousedown": [4, 7], "double_click": 4, "doubl": 4, "pos_from_time_callback": 4, "freq_assumpt": 4, "60": 4, "assumpt": 4, "mousemov": 4, "frequenc": 4, "accuraci": 4, "callabl": [4, 5, 6], "tupl": 4, "type_": 4, "timestamp": 4, "button": [4, 7, 8], "left": 4, "click_count": 4, "forc": 4, "tangential_pressur": 4, "tilt_x": 4, "tilt_i": 4, "twist": 4, "delta_x": 4, "delta_i": 4, "to_json": 4, "alt": 4, "shift": 4, "8": 4, "pointertyp": [4, 8], "mousepoint": 4, "pen": 4, "mousebutton": [4, 8], "middl": [4, 7], "right": 4, "16": 4, "eventtyp": [4, 8], "mousepress": 4, "mousereleas": 4, "wheel": 4, "mousewheel": 4, "import": [5, 8], "asyncio": [5, 8], "network_interceptor": 5, "networkinterceptor": 5, "interceptedrequest": 5, "interceptedauth": 5, "requestpattern": 5, "requeststag": 5, "def": [5, 8], "on_request": 5, "post": [5, 6], "continue_request": 5, "basic": 5, "intercept_respons": 5, "pattern": 5, "anyrequest": 5, "intercept_auth": 5, "interceptor": 5, "ensure_futur": 5, "isinst": 5, "virtual": 5, "zero": 5, "would": 5, "continue_auth": 5, "elif": 5, "stage": 5, "respons": [5, 6], "bodi": [5, 6, 8], "break": 5, "on_respons": 5, "on_auth": 5, "bypass_service_work": 5, "__init__": 5, "onrequest": 5, "onrespons": 5, "onauth": 5, "bypass": 5, "servic": 5, "worker": 5, "__aiter__": 5, "els": [5, 7], "possibl": [5, 7], "patter": 5, "is_redirect": 5, "redirect": [5, 6], "response_status_cod": 5, "301": 5, "302": 5, "303": 5, "307": 5, "308": 5, "requestpaus": 5, "frame_id": [5, 7], "requestid": [5, 6], "resource_typ": 5, "resourcetyp": 5, "network_id": 5, "requestwillbes": 5, "had": 5, "fire": 5, "response_error_reason": 5, "reason": 5, "response_head": 5, "header": [5, 6], "statu": [5, 6], "response_status_text": 5, "redirected_id": 5, "bypass_brows": 5, "allow_redirect": 5, "proxy_auth": 5, "extern": 5, "doe": 5, "tl": 5, "fingerprint": 5, "accordingli": 5, "technic": 5, "detect": [5, 7], "post_data": 5, "continu": 5, "fulfil": 5, "arrai": 5, "observ": 5, "overrid": 5, "continue_respons": 5, "binary_response_head": 5, "response_cod": 5, "response_phras": 5, "separ": 5, "seri": 5, "pair": 5, "treat": 5, "base64": [5, 6, 7], "encod": [5, 6], "phrase": 5, "ok": [5, 6], "resum": 5, "rais": 5, "cancel": 5, "done": 5, "fail_request": 5, "error_reason": 5, "fail": [5, 6, 7], "abort": 5, "timedout": 5, "accessdeni": 5, "connectionclos": 5, "connectionreset": 5, "connectionrefus": 5, "connectionabort": 5, "connectionfail": 5, "namenotresolv": 5, "internetdisconnect": 5, "addressunreach": 5, "blockedbycli": 5, "blockedbyrespons": 5, "authrequir": 5, "auth_challeng": 5, "authchalleng": 5, "cancelauth": 5, "providecredenti": 5, "challeng": 5, "digest": 5, "realm": 5, "pleas": 5, "fragment": 5, "url_frag": 5, "hash": 5, "present": 5, "has_post_data": 5, "postdata": 5, "still": [5, 7], "omit": 5, "flag": 5, "too": 5, "post_data_entri": 5, "postdataentri": 5, "mixed_content_typ": 5, "blockabl": 5, "mix": 5, "content": 5, "initial_prior": 5, "verylow": 5, "low": 5, "medium": 5, "high": [5, 6], "veryhigh": 5, "prioriti": [5, 6], "sent": 5, "referrer_polici": [5, 6], "unsaf": 5, "downgrad": 5, "polici": 5, "defin": 5, "tr": 5, "is_link_preload": 5, "via": 5, "link": [5, 6], "preload": 5, "trust_token_param": 5, "trusttokenparam": 5, "is_same_sit": 5, "consid": [5, 7], "enumer": 5, "urlpattern": 5, "anyrespons": 5, "url_pattern": 5, "request_stag": 5, "authalreadyhandledexcept": 5, "requestdoneexcept": 5, "is_remot": 6, "start_socket": 6, "usual": 6, "webwork": 6, "socket": 6, "singlecdpsocket": 6, "own": 6, "allow_not_on_map": 6, "keyboard": 6, "map": [6, 7], "unfocu": 6, "disabl": [6, 7], "get_histori": 6, "getnavigationhistori": 6, "navigationhistori": 6, "set_sourc": [6, 7], "15": [6, 8], "outerhtml": [6, 7], "dom": [6, 7], "try": 6, "framenavig": 6, "result": 6, "follow": 6, "chromedevtool": 6, "github": [6, 8], "io": 6, "empti": 6, "getresponsebodi": 6, "base64encod": 6, "credenti": 6, "cach": 6, "integr": 6, "keepal": 6, "mozilla": 6, "en": 6, "u": 6, "api": 6, "smth": 6, "status_cod": 6, "status_text": 6, "with_credenti": 6, "mime_typ": 6, "extra_head": 6, "put": 6, "includ": 6, "pars": 6, "extra": 6, "responsetext": 6, "responsetyp": 6, "responseurl": 6, "responsexml": 6, "statustext": 6, "responsehead": 6, "target_info": 6, "target_gett": 6, "info": 6, "itself": 6, "attach": [6, 7], "client": 6, "opener_id": 6, "open": 6, "can_access_open": 6, "opener_frame_id": 6, "browser_context_id": 6, "browsercontextid": 6, "subtyp": 6, "detail": [6, 7], "mai": 6, "portal": 6, "prerend": 6, "chromeinst": 6, "repres": [6, 7], "whole": 6, "unfortun": 6, "downloads_dir_for_context": 6, "isolated_exec_id": 7, "node_id": 7, "backend_node_id": 7, "is_ifram": 7, "interest": 7, "interact": 7, "through": 7, "interfac": 7, "fresh": 7, "check": 7, "valid": 7, "essenti": 7, "determin": 7, "If": 7, "staleelementreferenceexcept": 7, "thrown": 7, "futur": 7, "remoteobjectid": 7, "executioncontextid": 7, "nodeid": 7, "contentdocu": 7, "shadow_root": 7, "shadow": 7, "root": 7, "document_url": 7, "backendnodeid": 7, "classnam": 7, "idx": 7, "get_listen": 7, "domdebugg": 7, "geteventlisten": 7, "get_properti": 7, "elem": [7, 8], "tagnam": 7, "textcont": 7, "entri": 7, "highlight": 7, "fade": 7, "is_click": 7, "listener_depth": 7, "input": [7, 8], "select": 7, "textarea": 7, "video": 7, "wise": 7, "visible_timeout": 7, "spread_a": 7, "spread_b": 7, "bias_a": 7, "bias_b": 7, "border": 7, "05": 7, "scroll_to": 7, "ensure_click": 7, "abl": 7, "comput": 7, "box": 7, "model": 7, "spread": 7, "b": 7, "bia": 7, "minimum": 7, "toward": 7, "edg": 7, "rel": 7, "random": 7, "point": 7, "outsid": 7, "scroll": 7, "clickabl": 7, "reliabl": 7, "equival": 7, "6": 7, "std": 7, "99": 7, "7": 7, "write": 7, "click_on": 7, "insert": 7, "befor": 7, "set_fil": 7, "mid_loc": 7, "probabl": 7, "submit": 7, "form": 7, "It": 7, "dom_attribut": 7, "attribut": 7, "properli": [7, 8], "welel": 7, "get_dom_attribut": 7, "declar": 7, "markup": 7, "retriev": 7, "set_dom_attribut": 7, "get_attribut": 7, "is_select": 7, "checkbox": 7, "radio": 7, "is_en": 7, "is_displai": 7, "location_once_scrolled_into_view": 7, "rect": 7, "value_of_css_properti": 7, "property_nam": 7, "render": 7, "canva": 7, "box_model": 7, "boxmodel": 7, "aria_rol": 7, "aria": 7, "role": 7, "accessible_nam": 7, "level": 7, "screenshot_as_base64": 7, "screenshot_as_png": 7, "parent": 7, "python": 8, "pip": 8, "upgrad": 8, "selenium": 8, "chromeopt": 8, "nowsecur": 8, "nl": 8, "relax": 8, "domcontenteventfir": 8, "div": 8, "p": 8, "kaliiiiiiiiii": 8}, "objects": {"selenium_driverless.input.pointer": [[4, 0, 1, "", "BasePointer"], [4, 0, 1, "", "Buttons"], [4, 0, 1, "", "EventType"], [4, 0, 1, "", "Modifiers"], [4, 0, 1, "", "MouseButton"], [4, 0, 1, "", "Pointer"], [4, 0, 1, "", "PointerEvent"], [4, 0, 1, "", "PointerType"]], "selenium_driverless.input.pointer.BasePointer": [[4, 1, 1, "", "click"], [4, 1, 1, "", "dispatch"], [4, 1, 1, "", "double_click"], [4, 1, 1, "", "down"], [4, 1, 1, "", "move_path"], [4, 1, 1, "", "move_to"], [4, 1, 1, "", "up"]], "selenium_driverless.input.pointer.Buttons": [[4, 2, 1, "", "BACK"], [4, 2, 1, "", "DEFAULT"], [4, 2, 1, "", "FORWARD"], [4, 2, 1, "", "LEFT"], [4, 2, 1, "", "MIDDLE"], [4, 2, 1, "", "NONE"], [4, 2, 1, "", "RIGHT"]], "selenium_driverless.input.pointer.EventType": [[4, 2, 1, "", "MOVE"], [4, 2, 1, "", "PRESS"], [4, 2, 1, "", "RELEASE"], [4, 2, 1, "", "WHEEL"]], "selenium_driverless.input.pointer.Modifiers": [[4, 2, 1, "", "ALT"], [4, 2, 1, "", "COMMAND"], [4, 2, 1, "", "CTRL"], [4, 2, 1, "", "NONE"], [4, 2, 1, "", "SHIFT"]], "selenium_driverless.input.pointer.MouseButton": [[4, 2, 1, "", "BACK"], [4, 2, 1, "", "FORWARD"], [4, 2, 1, "", "LEFT"], [4, 2, 1, "", "MIDDLE"], [4, 2, 1, "", "NONE"], [4, 2, 1, "", "RIGHT"]], "selenium_driverless.input.pointer.Pointer": [[4, 1, 1, "", "click"], [4, 1, 1, "", "down"], [4, 1, 1, "", "move_to"], [4, 1, 1, "", "up"]], "selenium_driverless.input.pointer.PointerEvent": [[4, 1, 1, "", "to_json"]], "selenium_driverless.input.pointer.PointerType": [[4, 2, 1, "", "MOUSE"], [4, 2, 1, "", "PEN"]], "selenium_driverless.scripts.network_interceptor": [[5, 0, 1, "", "AuthAlreadyHandledException"], [5, 0, 1, "", "AuthChallenge"], [5, 0, 1, "", "InterceptedAuth"], [5, 0, 1, "", "InterceptedRequest"], [5, 0, 1, "", "NetworkInterceptor"], [5, 0, 1, "", "Request"], [5, 0, 1, "", "RequestDoneException"], [5, 0, 1, "", "RequestPattern"], [5, 0, 1, "", "RequestStages"]], "selenium_driverless.scripts.network_interceptor.AuthAlreadyHandledException": [[5, 3, 1, "", "request"]], "selenium_driverless.scripts.network_interceptor.AuthChallenge": [[5, 3, 1, "", "origin"], [5, 3, 1, "", "params"], [5, 3, 1, "", "realm"], [5, 3, 1, "", "scheme"], [5, 3, 1, "", "source"], [5, 3, 1, "", "target"]], "selenium_driverless.scripts.network_interceptor.InterceptedAuth": [[5, 3, 1, "", "auth_challenge"], [5, 1, 1, "", "cancel"], [5, 1, 1, "", "continue_auth"], [5, 3, 1, "", "frame_id"], [5, 3, 1, "", "id"], [5, 3, 1, "", "params"], [5, 3, 1, "", "request"], [5, 3, 1, "", "resource_type"], [5, 1, 1, "", "resume"], [5, 3, 1, "", "target"]], "selenium_driverless.scripts.network_interceptor.InterceptedRequest": [[5, 3, 1, "", "body"], [5, 1, 1, "", "bypass_browser"], [5, 1, 1, "", "continue_request"], [5, 1, 1, "", "continue_response"], [5, 1, 1, "", "fail_request"], [5, 3, 1, "", "frame_id"], [5, 1, 1, "", "fulfill"], [5, 3, 1, "", "id"], [5, 3, 1, "", "is_redirect"], [5, 3, 1, "", "network_id"], [5, 3, 1, "", "params"], [5, 3, 1, "", "redirected_id"], [5, 3, 1, "", "request"], [5, 3, 1, "", "resource_type"], [5, 3, 1, "", "response_error_reason"], [5, 3, 1, "", "response_headers"], [5, 3, 1, "", "response_status_code"], [5, 3, 1, "", "response_status_text"], [5, 1, 1, "", "resume"], [5, 3, 1, "", "stage"], [5, 3, 1, "", "target"]], "selenium_driverless.scripts.network_interceptor.NetworkInterceptor": [[5, 1, 1, "", "__aiter__"], [5, 1, 1, "", "__init__"], [5, 3, 1, "", "patterns"], [5, 3, 1, "", "target"]], "selenium_driverless.scripts.network_interceptor.Request": [[5, 3, 1, "", "has_post_data"], [5, 3, 1, "", "headers"], [5, 3, 1, "", "initial_priority"], [5, 3, 1, "", "is_link_preload"], [5, 3, 1, "", "is_same_site"], [5, 3, 1, "", "method"], [5, 3, 1, "", "mixed_content_type"], [5, 3, 1, "", "params"], [5, 3, 1, "", "post_data"], [5, 3, 1, "", "post_data_entries"], [5, 3, 1, "", "referrer_policy"], [5, 3, 1, "", "target"], [5, 3, 1, "", "trust_token_params"], [5, 3, 1, "", "url"], [5, 3, 1, "", "url_fragment"]], "selenium_driverless.scripts.network_interceptor.RequestDoneException": [[5, 3, 1, "", "request"]], "selenium_driverless.scripts.network_interceptor.RequestPattern": [[5, 2, 1, "", "AnyRequest"], [5, 2, 1, "", "AnyResponse"], [5, 1, 1, "", "new"]], "selenium_driverless.scripts.network_interceptor.RequestStages": [[5, 2, 1, "", "Request"], [5, 2, 1, "", "Response"]], "selenium_driverless.scripts.switch_to": [[1, 0, 1, "", "SwitchTo"]], "selenium_driverless.scripts.switch_to.SwitchTo": [[1, 3, 1, "", "alert"], [1, 1, 1, "", "default_content"], [1, 1, 1, "", "frame"], [1, 1, 1, "", "get_alert"], [1, 1, 1, "", "new_window"], [1, 1, 1, "", "target"], [1, 1, 1, "", "window"]], "selenium_driverless.types.base_target": [[6, 0, 1, "", "BaseTarget"]], "selenium_driverless.types.base_target.BaseTarget": [[6, 1, 1, "", "add_cdp_listener"], [6, 1, 1, "", "downloads_dir_for_context"], [6, 1, 1, "", "execute_cdp_cmd"], [6, 1, 1, "", "get_cdp_event_iter"], [6, 1, 1, "", "remove_cdp_listener"], [6, 3, 1, "", "socket"], [6, 1, 1, "", "wait_for_cdp"]], "selenium_driverless.types.by": [[0, 0, 1, "", "By"]], "selenium_driverless.types.by.By": [[0, 2, 1, "", "CLASS_NAME"], [0, 2, 1, "", "CSS"], [0, 2, 1, "", "CSS_SELECTOR"], [0, 2, 1, "", "ID"], [0, 2, 1, "", "NAME"], [0, 2, 1, "", "TAG_NAME"], [0, 2, 1, "", "XPATH"]], "selenium_driverless.types.context": [[3, 0, 1, "", "Context"]], "selenium_driverless.types.context.Context": [[3, 1, 1, "", "add_cdp_listener"], [3, 1, 1, "", "add_cookie"], [3, 1, 1, "", "back"], [3, 1, 1, "", "close"], [3, 3, 1, "", "current_pointer"], [3, 3, 1, "", "current_url"], [3, 3, 1, "", "current_window_handle"], [3, 1, 1, "", "delete_all_cookies"], [3, 1, 1, "", "delete_cookie"], [3, 1, 1, "", "delete_network_conditions"], [3, 3, 1, "", "downloads_dir"], [3, 1, 1, "", "eval_async"], [3, 1, 1, "", "execute_async_script"], [3, 1, 1, "", "execute_cdp_cmd"], [3, 1, 1, "", "execute_raw_script"], [3, 1, 1, "", "execute_script"], [3, 1, 1, "", "fetch"], [3, 1, 1, "", "find_element"], [3, 1, 1, "", "find_elements"], [3, 1, 1, "", "forward"], [3, 1, 1, "", "fullscreen_window"], [3, 1, 1, "", "get"], [3, 1, 1, "", "get_cdp_event_iter"], [3, 1, 1, "", "get_cookie"], [3, 1, 1, "", "get_cookies"], [3, 1, 1, "", "get_issue_message"], [3, 1, 1, "", "get_network_conditions"], [3, 1, 1, "", "get_screenshot_as_file"], [3, 1, 1, "", "get_screenshot_as_png"], [3, 1, 1, "", "get_sinks"], [3, 1, 1, "", "get_window_position"], [3, 1, 1, "", "get_window_rect"], [3, 1, 1, "", "get_window_size"], [3, 1, 1, "", "maximize_window"], [3, 1, 1, "", "minimize_window"], [3, 1, 1, "", "new_window"], [3, 3, 1, "", "page_source"], [3, 1, 1, "", "print_page"], [3, 1, 1, "", "quit"], [3, 1, 1, "", "refresh"], [3, 1, 1, "", "remove_cdp_listener"], [3, 1, 1, "", "save_screenshot"], [3, 1, 1, "", "save_snapshot"], [3, 1, 1, "", "search_elements"], [3, 1, 1, "", "send_keys"], [3, 1, 1, "", "set_download_behaviour"], [3, 1, 1, "", "set_network_conditions"], [3, 1, 1, "", "set_permissions"], [3, 1, 1, "", "set_sink_to_use"], [3, 1, 1, "", "set_window_position"], [3, 1, 1, "", "set_window_rect"], [3, 1, 1, "", "set_window_size"], [3, 1, 1, "", "snapshot"], [3, 1, 1, "", "start_desktop_mirroring"], [3, 1, 1, "", "start_tab_mirroring"], [3, 1, 1, "", "stop_casting"], [3, 3, 1, "", "switch_to"], [3, 3, 1, "", "title"], [3, 1, 1, "", "wait_download"], [3, 1, 1, "", "wait_for_cdp"], [3, 3, 1, "", "window_handles"], [3, 1, 1, "", "xhr"]], "selenium_driverless.types.options": [[2, 0, 1, "", "Options"]], "selenium_driverless.types.options.Options": [[2, 3, 1, "", "accept_insecure_certs"], [2, 1, 1, "", "add_argument"], [2, 1, 1, "", "add_arguments"], [2, 1, 1, "", "add_experimental_option"], [2, 1, 1, "", "add_extension"], [2, 3, 1, "", "arguments"], [2, 3, 1, "", "auto_clean_dirs"], [2, 3, 1, "", "binary_location"], [2, 3, 1, "", "debugger_address"], [2, 3, 1, "", "downloads_dir"], [2, 1, 1, "", "enable_mobile"], [2, 3, 1, "", "env"], [2, 3, 1, "", "headless"], [2, 1, 1, "", "ignore_local_proxy_environment_variables"], [2, 3, 1, "", "prefs"], [2, 3, 1, "", "single_proxy"], [2, 3, 1, "", "startup_url"], [2, 1, 1, "", "update_pref"], [2, 2, 1, "", "use_extension"], [2, 3, 1, "", "user_data_dir"]], "selenium_driverless.types.target": [[6, 0, 1, "", "Target"], [6, 0, 1, "", "TargetInfo"]], "selenium_driverless.types.target.Target": [[6, 1, 1, "", "activate"], [6, 1, 1, "", "add_cdp_listener"], [6, 1, 1, "", "add_cookie"], [6, 1, 1, "", "back"], [6, 1, 1, "", "close"], [6, 3, 1, "", "current_url"], [6, 1, 1, "", "delete_all_cookies"], [6, 1, 1, "", "delete_cookie"], [6, 1, 1, "", "delete_network_conditions"], [6, 1, 1, "", "eval_async"], [6, 1, 1, "", "execute_async_script"], [6, 1, 1, "", "execute_cdp_cmd"], [6, 1, 1, "", "execute_raw_script"], [6, 1, 1, "", "execute_script"], [6, 1, 1, "", "fetch"], [6, 1, 1, "", "find_element"], [6, 1, 1, "", "find_elements"], [6, 1, 1, "", "focus"], [6, 1, 1, "", "forward"], [6, 1, 1, "", "get"], [6, 1, 1, "", "get_cdp_event_iter"], [6, 1, 1, "", "get_cookie"], [6, 1, 1, "", "get_cookies"], [6, 1, 1, "", "get_history"], [6, 1, 1, "", "get_issue_message"], [6, 1, 1, "", "get_network_conditions"], [6, 1, 1, "", "get_screenshot_as_file"], [6, 1, 1, "", "get_screenshot_as_png"], [6, 1, 1, "", "get_sinks"], [6, 1, 1, "", "get_target_for_iframe"], [6, 1, 1, "", "get_targets_for_iframes"], [6, 3, 1, "", "page_source"], [6, 3, 1, "", "pointer"], [6, 1, 1, "", "print_page"], [6, 1, 1, "", "refresh"], [6, 1, 1, "", "remove_cdp_listener"], [6, 1, 1, "", "save_screenshot"], [6, 1, 1, "", "save_snapshot"], [6, 1, 1, "", "search_elements"], [6, 1, 1, "", "send_keys"], [6, 1, 1, "", "set_network_conditions"], [6, 1, 1, "", "set_sink_to_use"], [6, 1, 1, "", "set_source"], [6, 1, 1, "", "snapshot"], [6, 3, 1, "", "socket"], [6, 1, 1, "", "start_desktop_mirroring"], [6, 1, 1, "", "start_tab_mirroring"], [6, 1, 1, "", "stop_casting"], [6, 3, 1, "", "title"], [6, 1, 1, "", "unfocus"], [6, 1, 1, "", "wait_download"], [6, 1, 1, "", "wait_for_cdp"], [6, 1, 1, "", "xhr"]], "selenium_driverless.types.target.TargetInfo": [[6, 3, 1, "", "Target"], [6, 3, 1, "", "attached"], [6, 3, 1, "", "browser_context_id"], [6, 3, 1, "", "can_access_opener"], [6, 3, 1, "", "id"], [6, 3, 1, "", "opener_frame_id"], [6, 3, 1, "", "opener_id"], [6, 3, 1, "", "subtype"]], "selenium_driverless.types.webelement": [[7, 0, 1, "", "WebElement"]], "selenium_driverless.types.webelement.WebElement": [[7, 3, 1, "", "accessible_name"], [7, 3, 1, "", "aria_role"], [7, 3, 1, "", "backend_node_id"], [7, 3, 1, "", "box_model"], [7, 3, 1, "", "class_name"], [7, 1, 1, "", "clear"], [7, 1, 1, "", "click"], [7, 3, 1, "", "content_document"], [7, 3, 1, "", "context_id"], [7, 3, 1, "", "document_url"], [7, 3, 1, "", "dom_attributes"], [7, 1, 1, "", "eval_async"], [7, 1, 1, "", "execute_async_script"], [7, 1, 1, "", "execute_script"], [7, 1, 1, "", "find_element"], [7, 1, 1, "", "find_elements"], [7, 1, 1, "", "focus"], [7, 1, 1, "", "get_attribute"], [7, 1, 1, "", "get_dom_attribute"], [7, 1, 1, "", "get_listeners"], [7, 1, 1, "", "get_property"], [7, 1, 1, "", "highlight"], [7, 1, 1, "", "is_clickable"], [7, 1, 1, "", "is_displayed"], [7, 1, 1, "", "is_enabled"], [7, 1, 1, "", "is_selected"], [7, 3, 1, "", "location"], [7, 3, 1, "", "location_once_scrolled_into_view"], [7, 1, 1, "", "mid_location"], [7, 3, 1, "", "node_id"], [7, 3, 1, "", "obj_id"], [7, 3, 1, "", "parent"], [7, 3, 1, "", "rect"], [7, 1, 1, "", "remove"], [7, 1, 1, "", "screenshot"], [7, 3, 1, "", "screenshot_as_base64"], [7, 3, 1, "", "screenshot_as_png"], [7, 1, 1, "", "scroll_to"], [7, 1, 1, "", "send_keys"], [7, 1, 1, "", "set_dom_attribute"], [7, 1, 1, "", "set_file"], [7, 1, 1, "", "set_files"], [7, 1, 1, "", "set_source"], [7, 3, 1, "", "shadow_root"], [7, 3, 1, "", "shadow_roots"], [7, 3, 1, "", "size"], [7, 3, 1, "", "source"], [7, 1, 1, "", "submit"], [7, 3, 1, "", "tag_name"], [7, 3, 1, "", "text"], [7, 3, 1, "", "value"], [7, 1, 1, "", "value_of_css_property"], [7, 1, 1, "", "write"]], "selenium_driverless.webdriver": [[1, 0, 1, "", "Chrome"]], "selenium_driverless.webdriver.Chrome": [[1, 1, 1, "", "add_cdp_listener"], [1, 1, 1, "", "add_cookie"], [1, 1, 1, "", "back"], [1, 3, 1, "", "base_context"], [1, 3, 1, "", "base_target"], [1, 1, 1, "", "clear_auth"], [1, 1, 1, "", "clear_proxy"], [1, 1, 1, "", "close"], [1, 3, 1, "", "contexts"], [1, 3, 1, "", "current_context"], [1, 3, 1, "", "current_pointer"], [1, 3, 1, "", "current_target"], [1, 3, 1, "", "current_target_info"], [1, 3, 1, "", "current_url"], [1, 3, 1, "", "current_window_handle"], [1, 3, 1, "", "current_window_id"], [1, 1, 1, "", "delete_all_cookies"], [1, 1, 1, "", "delete_cookie"], [1, 1, 1, "", "delete_network_conditions"], [1, 3, 1, "", "downloads_dir"], [1, 1, 1, "", "ensure_extensions_incognito_allowed"], [1, 1, 1, "", "eval_async"], [1, 1, 1, "", "execute_async_script"], [1, 1, 1, "", "execute_cdp_cmd"], [1, 1, 1, "", "execute_raw_script"], [1, 1, 1, "", "execute_script"], [1, 1, 1, "", "fetch"], [1, 1, 1, "", "find_element"], [1, 1, 1, "", "find_elements"], [1, 1, 1, "", "focus"], [1, 1, 1, "", "forward"], [1, 3, 1, "", "frame_tree"], [1, 1, 1, "", "fullscreen_window"], [1, 1, 1, "", "get"], [1, 1, 1, "", "get_cdp_event_iter"], [1, 1, 1, "", "get_cookie"], [1, 1, 1, "", "get_cookies"], [1, 1, 1, "", "get_issue_message"], [1, 1, 1, "", "get_network_conditions"], [1, 1, 1, "", "get_screenshot_as_file"], [1, 1, 1, "", "get_screenshot_as_png"], [1, 1, 1, "", "get_sinks"], [1, 1, 1, "", "get_target"], [1, 1, 1, "", "get_target_for_iframe"], [1, 1, 1, "", "get_targets"], [1, 1, 1, "", "get_targets_for_iframes"], [1, 1, 1, "", "get_window_position"], [1, 1, 1, "", "get_window_rect"], [1, 1, 1, "", "get_window_size"], [1, 1, 1, "", "maximize_window"], [1, 1, 1, "", "minimize_window"], [1, 3, 1, "", "mv3_extension"], [1, 1, 1, "", "new_context"], [1, 1, 1, "", "new_window"], [1, 1, 1, "", "normalize_window"], [1, 3, 1, "", "page_source"], [1, 1, 1, "", "print_page"], [1, 1, 1, "", "quit"], [1, 1, 1, "", "refresh"], [1, 1, 1, "", "remove_cdp_listener"], [1, 1, 1, "", "save_screenshot"], [1, 1, 1, "", "save_snapshot"], [1, 1, 1, "", "search_elements"], [1, 1, 1, "", "send_keys"], [1, 1, 1, "", "set_auth"], [1, 1, 1, "", "set_download_behaviour"], [1, 1, 1, "", "set_network_conditions"], [1, 1, 1, "", "set_permissions"], [1, 1, 1, "", "set_proxy"], [1, 1, 1, "", "set_single_proxy"], [1, 1, 1, "", "set_sink_to_use"], [1, 1, 1, "", "set_window_position"], [1, 1, 1, "", "set_window_rect"], [1, 1, 1, "", "set_window_size"], [1, 1, 1, "", "set_window_state"], [1, 1, 1, "", "sleep"], [1, 1, 1, "", "snapshot"], [1, 1, 1, "", "start_desktop_mirroring"], [1, 1, 1, "", "start_tab_mirroring"], [1, 1, 1, "", "stop_casting"], [1, 3, 1, "", "switch_to"], [1, 3, 1, "", "targets"], [1, 3, 1, "", "title"], [1, 1, 1, "", "wait_download"], [1, 1, 1, "", "wait_for_cdp"], [1, 3, 1, "", "window_handles"], [1, 1, 1, "", "xhr"]]}, "objtypes": {"0": "py:class", "1": "py:method", "2": "py:attribute", "3": "py:property"}, "objnames": {"0": ["py", "class", "Python class"], "1": ["py", "method", "Python method"], "2": ["py", "attribute", "Python attribute"], "3": ["py", "property", "Python property"]}, "titleterms": {"By": 0, "element": 0, "locat": 0, "webdriv": 1, "chrome": 1, "chromeopt": 2, "context": 3, "input": 4, "request": 5, "intercept": 5, "exampl": 5, "script": 5, "api": [5, 8], "target": 6, "webel": 7, "document": 8, "driverless": 8, "instal": 8, "usag": 8, "sourc": 8}, "envversion": {"sphinx.domains.c": 3, "sphinx.domains.changeset": 1, "sphinx.domains.citation": 1, "sphinx.domains.cpp": 9, "sphinx.domains.index": 1, "sphinx.domains.javascript": 3, "sphinx.domains.math": 2, "sphinx.domains.python": 4, "sphinx.domains.rst": 2, "sphinx.domains.std": 2, "sphinx.ext.viewcode": 1, "sphinx": 58}, "alltitles": {"By Element Locator": [[0, "by-element-locator"]], "ChromeOptions": [[2, "chromeoptions"]], "Input": [[4, "input"]], "Request-Interception": [[5, "request-interception"]], "Example Script": [[5, "example-script"]], "API": [[5, "api"], [8, "api"]], "WebElement": [[7, "webelement"]], "Documentation of Driverless": [[8, "documentation-of-driverless"]], "Installation": [[8, "installation"]], "Usage": [[8, "usage"]], "Source": [[8, "source"]], "webdriver.Chrome": [[1, "webdriver-chrome"]], "Context": [[3, "context"]], "Target": [[6, "target"]]}, "indexentries": {"chrome (class in selenium_driverless.webdriver)": [[1, "selenium_driverless.webdriver.Chrome"]], "switchto (class in selenium_driverless.scripts.switch_to)": [[1, "selenium_driverless.scripts.switch_to.SwitchTo"]], "add_cdp_listener() (selenium_driverless.webdriver.chrome method)": [[1, "selenium_driverless.webdriver.Chrome.add_cdp_listener"]], "add_cookie() (selenium_driverless.webdriver.chrome method)": [[1, "selenium_driverless.webdriver.Chrome.add_cookie"]], "alert (selenium_driverless.scripts.switch_to.switchto property)": [[1, "selenium_driverless.scripts.switch_to.SwitchTo.alert"]], "back() (selenium_driverless.webdriver.chrome method)": [[1, "selenium_driverless.webdriver.Chrome.back"]], "base_context (selenium_driverless.webdriver.chrome property)": [[1, "selenium_driverless.webdriver.Chrome.base_context"]], "base_target (selenium_driverless.webdriver.chrome property)": [[1, "selenium_driverless.webdriver.Chrome.base_target"]], "clear_auth() (selenium_driverless.webdriver.chrome method)": [[1, "selenium_driverless.webdriver.Chrome.clear_auth"]], "clear_proxy() (selenium_driverless.webdriver.chrome method)": [[1, "selenium_driverless.webdriver.Chrome.clear_proxy"]], "close() (selenium_driverless.webdriver.chrome method)": [[1, "selenium_driverless.webdriver.Chrome.close"]], "contexts (selenium_driverless.webdriver.chrome property)": [[1, "selenium_driverless.webdriver.Chrome.contexts"]], "current_context (selenium_driverless.webdriver.chrome property)": [[1, "selenium_driverless.webdriver.Chrome.current_context"]], "current_pointer (selenium_driverless.webdriver.chrome property)": [[1, "selenium_driverless.webdriver.Chrome.current_pointer"]], "current_target (selenium_driverless.webdriver.chrome property)": [[1, "selenium_driverless.webdriver.Chrome.current_target"]], "current_target_info (selenium_driverless.webdriver.chrome property)": [[1, "selenium_driverless.webdriver.Chrome.current_target_info"]], "current_url (selenium_driverless.webdriver.chrome property)": [[1, "selenium_driverless.webdriver.Chrome.current_url"]], "current_window_handle (selenium_driverless.webdriver.chrome property)": [[1, "selenium_driverless.webdriver.Chrome.current_window_handle"]], "current_window_id (selenium_driverless.webdriver.chrome property)": [[1, "selenium_driverless.webdriver.Chrome.current_window_id"]], "default_content() (selenium_driverless.scripts.switch_to.switchto method)": [[1, "selenium_driverless.scripts.switch_to.SwitchTo.default_content"]], "delete_all_cookies() (selenium_driverless.webdriver.chrome method)": [[1, "selenium_driverless.webdriver.Chrome.delete_all_cookies"]], "delete_cookie() (selenium_driverless.webdriver.chrome method)": [[1, "selenium_driverless.webdriver.Chrome.delete_cookie"]], "delete_network_conditions() (selenium_driverless.webdriver.chrome method)": [[1, "selenium_driverless.webdriver.Chrome.delete_network_conditions"]], "downloads_dir (selenium_driverless.webdriver.chrome property)": [[1, "selenium_driverless.webdriver.Chrome.downloads_dir"]], "ensure_extensions_incognito_allowed() (selenium_driverless.webdriver.chrome method)": [[1, "selenium_driverless.webdriver.Chrome.ensure_extensions_incognito_allowed"]], "eval_async() (selenium_driverless.webdriver.chrome method)": [[1, "selenium_driverless.webdriver.Chrome.eval_async"]], "execute_async_script() (selenium_driverless.webdriver.chrome method)": [[1, "selenium_driverless.webdriver.Chrome.execute_async_script"]], "execute_cdp_cmd() (selenium_driverless.webdriver.chrome method)": [[1, "selenium_driverless.webdriver.Chrome.execute_cdp_cmd"]], "execute_raw_script() (selenium_driverless.webdriver.chrome method)": [[1, "selenium_driverless.webdriver.Chrome.execute_raw_script"]], "execute_script() (selenium_driverless.webdriver.chrome method)": [[1, "selenium_driverless.webdriver.Chrome.execute_script"]], "fetch() (selenium_driverless.webdriver.chrome method)": [[1, "selenium_driverless.webdriver.Chrome.fetch"]], "find_element() (selenium_driverless.webdriver.chrome method)": [[1, "selenium_driverless.webdriver.Chrome.find_element"]], "find_elements() (selenium_driverless.webdriver.chrome method)": [[1, "selenium_driverless.webdriver.Chrome.find_elements"]], "focus() (selenium_driverless.webdriver.chrome method)": [[1, "selenium_driverless.webdriver.Chrome.focus"]], "forward() (selenium_driverless.webdriver.chrome method)": [[1, "selenium_driverless.webdriver.Chrome.forward"]], "frame() (selenium_driverless.scripts.switch_to.switchto method)": [[1, "selenium_driverless.scripts.switch_to.SwitchTo.frame"]], "frame_tree (selenium_driverless.webdriver.chrome property)": [[1, "selenium_driverless.webdriver.Chrome.frame_tree"]], "fullscreen_window() (selenium_driverless.webdriver.chrome method)": [[1, "selenium_driverless.webdriver.Chrome.fullscreen_window"]], "get() (selenium_driverless.webdriver.chrome method)": [[1, "selenium_driverless.webdriver.Chrome.get"]], "get_alert() (selenium_driverless.scripts.switch_to.switchto method)": [[1, "selenium_driverless.scripts.switch_to.SwitchTo.get_alert"]], "get_cdp_event_iter() (selenium_driverless.webdriver.chrome method)": [[1, "selenium_driverless.webdriver.Chrome.get_cdp_event_iter"]], "get_cookie() (selenium_driverless.webdriver.chrome method)": [[1, "selenium_driverless.webdriver.Chrome.get_cookie"]], "get_cookies() (selenium_driverless.webdriver.chrome method)": [[1, "selenium_driverless.webdriver.Chrome.get_cookies"]], "get_issue_message() (selenium_driverless.webdriver.chrome method)": [[1, "selenium_driverless.webdriver.Chrome.get_issue_message"]], "get_network_conditions() (selenium_driverless.webdriver.chrome method)": [[1, "selenium_driverless.webdriver.Chrome.get_network_conditions"]], "get_screenshot_as_file() (selenium_driverless.webdriver.chrome method)": [[1, "selenium_driverless.webdriver.Chrome.get_screenshot_as_file"]], "get_screenshot_as_png() (selenium_driverless.webdriver.chrome method)": [[1, "selenium_driverless.webdriver.Chrome.get_screenshot_as_png"]], "get_sinks() (selenium_driverless.webdriver.chrome method)": [[1, "selenium_driverless.webdriver.Chrome.get_sinks"]], "get_target() (selenium_driverless.webdriver.chrome method)": [[1, "selenium_driverless.webdriver.Chrome.get_target"]], "get_target_for_iframe() (selenium_driverless.webdriver.chrome method)": [[1, "selenium_driverless.webdriver.Chrome.get_target_for_iframe"]], "get_targets() (selenium_driverless.webdriver.chrome method)": [[1, "selenium_driverless.webdriver.Chrome.get_targets"]], "get_targets_for_iframes() (selenium_driverless.webdriver.chrome method)": [[1, "selenium_driverless.webdriver.Chrome.get_targets_for_iframes"]], "get_window_position() (selenium_driverless.webdriver.chrome method)": [[1, "selenium_driverless.webdriver.Chrome.get_window_position"]], "get_window_rect() (selenium_driverless.webdriver.chrome method)": [[1, "selenium_driverless.webdriver.Chrome.get_window_rect"]], "get_window_size() (selenium_driverless.webdriver.chrome method)": [[1, "selenium_driverless.webdriver.Chrome.get_window_size"]], "maximize_window() (selenium_driverless.webdriver.chrome method)": [[1, "selenium_driverless.webdriver.Chrome.maximize_window"]], "minimize_window() (selenium_driverless.webdriver.chrome method)": [[1, "selenium_driverless.webdriver.Chrome.minimize_window"]], "mv3_extension (selenium_driverless.webdriver.chrome property)": [[1, "selenium_driverless.webdriver.Chrome.mv3_extension"]], "new_context() (selenium_driverless.webdriver.chrome method)": [[1, "selenium_driverless.webdriver.Chrome.new_context"]], "new_window() (selenium_driverless.scripts.switch_to.switchto method)": [[1, "selenium_driverless.scripts.switch_to.SwitchTo.new_window"]], "new_window() (selenium_driverless.webdriver.chrome method)": [[1, "selenium_driverless.webdriver.Chrome.new_window"]], "normalize_window() (selenium_driverless.webdriver.chrome method)": [[1, "selenium_driverless.webdriver.Chrome.normalize_window"]], "page_source (selenium_driverless.webdriver.chrome property)": [[1, "selenium_driverless.webdriver.Chrome.page_source"]], "print_page() (selenium_driverless.webdriver.chrome method)": [[1, "selenium_driverless.webdriver.Chrome.print_page"]], "quit() (selenium_driverless.webdriver.chrome method)": [[1, "selenium_driverless.webdriver.Chrome.quit"]], "refresh() (selenium_driverless.webdriver.chrome method)": [[1, "selenium_driverless.webdriver.Chrome.refresh"]], "remove_cdp_listener() (selenium_driverless.webdriver.chrome method)": [[1, "selenium_driverless.webdriver.Chrome.remove_cdp_listener"]], "save_screenshot() (selenium_driverless.webdriver.chrome method)": [[1, "selenium_driverless.webdriver.Chrome.save_screenshot"]], "save_snapshot() (selenium_driverless.webdriver.chrome method)": [[1, "selenium_driverless.webdriver.Chrome.save_snapshot"]], "search_elements() (selenium_driverless.webdriver.chrome method)": [[1, "selenium_driverless.webdriver.Chrome.search_elements"]], "send_keys() (selenium_driverless.webdriver.chrome method)": [[1, "selenium_driverless.webdriver.Chrome.send_keys"]], "set_auth() (selenium_driverless.webdriver.chrome method)": [[1, "selenium_driverless.webdriver.Chrome.set_auth"]], "set_download_behaviour() (selenium_driverless.webdriver.chrome method)": [[1, "selenium_driverless.webdriver.Chrome.set_download_behaviour"]], "set_network_conditions() (selenium_driverless.webdriver.chrome method)": [[1, "selenium_driverless.webdriver.Chrome.set_network_conditions"]], "set_permissions() (selenium_driverless.webdriver.chrome method)": [[1, "selenium_driverless.webdriver.Chrome.set_permissions"]], "set_proxy() (selenium_driverless.webdriver.chrome method)": [[1, "selenium_driverless.webdriver.Chrome.set_proxy"]], "set_single_proxy() (selenium_driverless.webdriver.chrome method)": [[1, "selenium_driverless.webdriver.Chrome.set_single_proxy"]], "set_sink_to_use() (selenium_driverless.webdriver.chrome method)": [[1, "selenium_driverless.webdriver.Chrome.set_sink_to_use"]], "set_window_position() (selenium_driverless.webdriver.chrome method)": [[1, "selenium_driverless.webdriver.Chrome.set_window_position"]], "set_window_rect() (selenium_driverless.webdriver.chrome method)": [[1, "selenium_driverless.webdriver.Chrome.set_window_rect"]], "set_window_size() (selenium_driverless.webdriver.chrome method)": [[1, "selenium_driverless.webdriver.Chrome.set_window_size"]], "set_window_state() (selenium_driverless.webdriver.chrome method)": [[1, "selenium_driverless.webdriver.Chrome.set_window_state"]], "sleep() (selenium_driverless.webdriver.chrome static method)": [[1, "selenium_driverless.webdriver.Chrome.sleep"]], "snapshot() (selenium_driverless.webdriver.chrome method)": [[1, "selenium_driverless.webdriver.Chrome.snapshot"]], "start_desktop_mirroring() (selenium_driverless.webdriver.chrome method)": [[1, "selenium_driverless.webdriver.Chrome.start_desktop_mirroring"]], "start_tab_mirroring() (selenium_driverless.webdriver.chrome method)": [[1, "selenium_driverless.webdriver.Chrome.start_tab_mirroring"]], "stop_casting() (selenium_driverless.webdriver.chrome method)": [[1, "selenium_driverless.webdriver.Chrome.stop_casting"]], "switch_to (selenium_driverless.webdriver.chrome property)": [[1, "selenium_driverless.webdriver.Chrome.switch_to"]], "target() (selenium_driverless.scripts.switch_to.switchto method)": [[1, "selenium_driverless.scripts.switch_to.SwitchTo.target"]], "targets (selenium_driverless.webdriver.chrome property)": [[1, "selenium_driverless.webdriver.Chrome.targets"]], "title (selenium_driverless.webdriver.chrome property)": [[1, "selenium_driverless.webdriver.Chrome.title"]], "wait_download() (selenium_driverless.webdriver.chrome method)": [[1, "selenium_driverless.webdriver.Chrome.wait_download"]], "wait_for_cdp() (selenium_driverless.webdriver.chrome method)": [[1, "selenium_driverless.webdriver.Chrome.wait_for_cdp"]], "window() (selenium_driverless.scripts.switch_to.switchto method)": [[1, "selenium_driverless.scripts.switch_to.SwitchTo.window"]], "window_handles (selenium_driverless.webdriver.chrome property)": [[1, "selenium_driverless.webdriver.Chrome.window_handles"]], "xhr() (selenium_driverless.webdriver.chrome method)": [[1, "selenium_driverless.webdriver.Chrome.xhr"]], "context (class in selenium_driverless.types.context)": [[3, "selenium_driverless.types.context.Context"]], "add_cdp_listener() (selenium_driverless.types.context.context method)": [[3, "selenium_driverless.types.context.Context.add_cdp_listener"]], "add_cookie() (selenium_driverless.types.context.context method)": [[3, "selenium_driverless.types.context.Context.add_cookie"]], "back() (selenium_driverless.types.context.context method)": [[3, "selenium_driverless.types.context.Context.back"]], "close() (selenium_driverless.types.context.context method)": [[3, "selenium_driverless.types.context.Context.close"]], "current_pointer (selenium_driverless.types.context.context property)": [[3, "selenium_driverless.types.context.Context.current_pointer"]], "current_url (selenium_driverless.types.context.context property)": [[3, "selenium_driverless.types.context.Context.current_url"]], "current_window_handle (selenium_driverless.types.context.context property)": [[3, "selenium_driverless.types.context.Context.current_window_handle"]], "delete_all_cookies() (selenium_driverless.types.context.context method)": [[3, "selenium_driverless.types.context.Context.delete_all_cookies"]], "delete_cookie() (selenium_driverless.types.context.context method)": [[3, "selenium_driverless.types.context.Context.delete_cookie"]], "delete_network_conditions() (selenium_driverless.types.context.context method)": [[3, "selenium_driverless.types.context.Context.delete_network_conditions"]], "downloads_dir (selenium_driverless.types.context.context property)": [[3, "selenium_driverless.types.context.Context.downloads_dir"]], "eval_async() (selenium_driverless.types.context.context method)": [[3, "selenium_driverless.types.context.Context.eval_async"]], "execute_async_script() (selenium_driverless.types.context.context method)": [[3, "selenium_driverless.types.context.Context.execute_async_script"]], "execute_cdp_cmd() (selenium_driverless.types.context.context method)": [[3, "selenium_driverless.types.context.Context.execute_cdp_cmd"]], "execute_raw_script() (selenium_driverless.types.context.context method)": [[3, "selenium_driverless.types.context.Context.execute_raw_script"]], "execute_script() (selenium_driverless.types.context.context method)": [[3, "selenium_driverless.types.context.Context.execute_script"]], "fetch() (selenium_driverless.types.context.context method)": [[3, "selenium_driverless.types.context.Context.fetch"]], "find_element() (selenium_driverless.types.context.context method)": [[3, "selenium_driverless.types.context.Context.find_element"]], "find_elements() (selenium_driverless.types.context.context method)": [[3, "selenium_driverless.types.context.Context.find_elements"]], "forward() (selenium_driverless.types.context.context method)": [[3, "selenium_driverless.types.context.Context.forward"]], "fullscreen_window() (selenium_driverless.types.context.context method)": [[3, "selenium_driverless.types.context.Context.fullscreen_window"]], "get() (selenium_driverless.types.context.context method)": [[3, "selenium_driverless.types.context.Context.get"]], "get_cdp_event_iter() (selenium_driverless.types.context.context method)": [[3, "selenium_driverless.types.context.Context.get_cdp_event_iter"]], "get_cookie() (selenium_driverless.types.context.context method)": [[3, "selenium_driverless.types.context.Context.get_cookie"]], "get_cookies() (selenium_driverless.types.context.context method)": [[3, "selenium_driverless.types.context.Context.get_cookies"]], "get_issue_message() (selenium_driverless.types.context.context method)": [[3, "selenium_driverless.types.context.Context.get_issue_message"]], "get_network_conditions() (selenium_driverless.types.context.context method)": [[3, "selenium_driverless.types.context.Context.get_network_conditions"]], "get_screenshot_as_file() (selenium_driverless.types.context.context method)": [[3, "selenium_driverless.types.context.Context.get_screenshot_as_file"]], "get_screenshot_as_png() (selenium_driverless.types.context.context method)": [[3, "selenium_driverless.types.context.Context.get_screenshot_as_png"]], "get_sinks() (selenium_driverless.types.context.context method)": [[3, "selenium_driverless.types.context.Context.get_sinks"]], "get_window_position() (selenium_driverless.types.context.context method)": [[3, "selenium_driverless.types.context.Context.get_window_position"]], "get_window_rect() (selenium_driverless.types.context.context method)": [[3, "selenium_driverless.types.context.Context.get_window_rect"]], "get_window_size() (selenium_driverless.types.context.context method)": [[3, "selenium_driverless.types.context.Context.get_window_size"]], "maximize_window() (selenium_driverless.types.context.context method)": [[3, "selenium_driverless.types.context.Context.maximize_window"]], "minimize_window() (selenium_driverless.types.context.context method)": [[3, "selenium_driverless.types.context.Context.minimize_window"]], "new_window() (selenium_driverless.types.context.context method)": [[3, "selenium_driverless.types.context.Context.new_window"]], "page_source (selenium_driverless.types.context.context property)": [[3, "selenium_driverless.types.context.Context.page_source"]], "print_page() (selenium_driverless.types.context.context method)": [[3, "selenium_driverless.types.context.Context.print_page"]], "quit() (selenium_driverless.types.context.context method)": [[3, "selenium_driverless.types.context.Context.quit"]], "refresh() (selenium_driverless.types.context.context method)": [[3, "selenium_driverless.types.context.Context.refresh"]], "remove_cdp_listener() (selenium_driverless.types.context.context method)": [[3, "selenium_driverless.types.context.Context.remove_cdp_listener"]], "save_screenshot() (selenium_driverless.types.context.context method)": [[3, "selenium_driverless.types.context.Context.save_screenshot"]], "save_snapshot() (selenium_driverless.types.context.context method)": [[3, "selenium_driverless.types.context.Context.save_snapshot"]], "search_elements() (selenium_driverless.types.context.context method)": [[3, "selenium_driverless.types.context.Context.search_elements"]], "send_keys() (selenium_driverless.types.context.context method)": [[3, "selenium_driverless.types.context.Context.send_keys"]], "set_download_behaviour() (selenium_driverless.types.context.context method)": [[3, "selenium_driverless.types.context.Context.set_download_behaviour"]], "set_network_conditions() (selenium_driverless.types.context.context method)": [[3, "selenium_driverless.types.context.Context.set_network_conditions"]], "set_permissions() (selenium_driverless.types.context.context method)": [[3, "selenium_driverless.types.context.Context.set_permissions"]], "set_sink_to_use() (selenium_driverless.types.context.context method)": [[3, "selenium_driverless.types.context.Context.set_sink_to_use"]], "set_window_position() (selenium_driverless.types.context.context method)": [[3, "selenium_driverless.types.context.Context.set_window_position"]], "set_window_rect() (selenium_driverless.types.context.context method)": [[3, "selenium_driverless.types.context.Context.set_window_rect"]], "set_window_size() (selenium_driverless.types.context.context method)": [[3, "selenium_driverless.types.context.Context.set_window_size"]], "snapshot() (selenium_driverless.types.context.context method)": [[3, "selenium_driverless.types.context.Context.snapshot"]], "start_desktop_mirroring() (selenium_driverless.types.context.context method)": [[3, "selenium_driverless.types.context.Context.start_desktop_mirroring"]], "start_tab_mirroring() (selenium_driverless.types.context.context method)": [[3, "selenium_driverless.types.context.Context.start_tab_mirroring"]], "stop_casting() (selenium_driverless.types.context.context method)": [[3, "selenium_driverless.types.context.Context.stop_casting"]], "switch_to (selenium_driverless.types.context.context property)": [[3, "selenium_driverless.types.context.Context.switch_to"]], "title (selenium_driverless.types.context.context property)": [[3, "selenium_driverless.types.context.Context.title"]], "wait_download() (selenium_driverless.types.context.context method)": [[3, "selenium_driverless.types.context.Context.wait_download"]], "wait_for_cdp() (selenium_driverless.types.context.context method)": [[3, "selenium_driverless.types.context.Context.wait_for_cdp"]], "window_handles (selenium_driverless.types.context.context property)": [[3, "selenium_driverless.types.context.Context.window_handles"]], "xhr() (selenium_driverless.types.context.context method)": [[3, "selenium_driverless.types.context.Context.xhr"]], "basetarget (class in selenium_driverless.types.base_target)": [[6, "selenium_driverless.types.base_target.BaseTarget"]], "target (class in selenium_driverless.types.target)": [[6, "selenium_driverless.types.target.Target"]], "target (selenium_driverless.types.target.targetinfo property)": [[6, "selenium_driverless.types.target.TargetInfo.Target"]], "targetinfo (class in selenium_driverless.types.target)": [[6, "selenium_driverless.types.target.TargetInfo"]], "activate() (selenium_driverless.types.target.target method)": [[6, "selenium_driverless.types.target.Target.activate"]], "add_cdp_listener() (selenium_driverless.types.base_target.basetarget method)": [[6, "selenium_driverless.types.base_target.BaseTarget.add_cdp_listener"]], "add_cdp_listener() (selenium_driverless.types.target.target method)": [[6, "selenium_driverless.types.target.Target.add_cdp_listener"]], "add_cookie() (selenium_driverless.types.target.target method)": [[6, "selenium_driverless.types.target.Target.add_cookie"]], "attached (selenium_driverless.types.target.targetinfo property)": [[6, "selenium_driverless.types.target.TargetInfo.attached"]], "back() (selenium_driverless.types.target.target method)": [[6, "selenium_driverless.types.target.Target.back"]], "browser_context_id (selenium_driverless.types.target.targetinfo property)": [[6, "selenium_driverless.types.target.TargetInfo.browser_context_id"]], "can_access_opener (selenium_driverless.types.target.targetinfo property)": [[6, "selenium_driverless.types.target.TargetInfo.can_access_opener"]], "close() (selenium_driverless.types.target.target method)": [[6, "selenium_driverless.types.target.Target.close"]], "current_url (selenium_driverless.types.target.target property)": [[6, "selenium_driverless.types.target.Target.current_url"]], "delete_all_cookies() (selenium_driverless.types.target.target method)": [[6, "selenium_driverless.types.target.Target.delete_all_cookies"]], "delete_cookie() (selenium_driverless.types.target.target method)": [[6, "selenium_driverless.types.target.Target.delete_cookie"]], "delete_network_conditions() (selenium_driverless.types.target.target method)": [[6, "selenium_driverless.types.target.Target.delete_network_conditions"]], "downloads_dir_for_context() (selenium_driverless.types.base_target.basetarget method)": [[6, "selenium_driverless.types.base_target.BaseTarget.downloads_dir_for_context"]], "eval_async() (selenium_driverless.types.target.target method)": [[6, "selenium_driverless.types.target.Target.eval_async"]], "execute_async_script() (selenium_driverless.types.target.target method)": [[6, "selenium_driverless.types.target.Target.execute_async_script"]], "execute_cdp_cmd() (selenium_driverless.types.base_target.basetarget method)": [[6, "selenium_driverless.types.base_target.BaseTarget.execute_cdp_cmd"]], "execute_cdp_cmd() (selenium_driverless.types.target.target method)": [[6, "selenium_driverless.types.target.Target.execute_cdp_cmd"]], "execute_raw_script() (selenium_driverless.types.target.target method)": [[6, "selenium_driverless.types.target.Target.execute_raw_script"]], "execute_script() (selenium_driverless.types.target.target method)": [[6, "selenium_driverless.types.target.Target.execute_script"]], "fetch() (selenium_driverless.types.target.target method)": [[6, "selenium_driverless.types.target.Target.fetch"]], "find_element() (selenium_driverless.types.target.target method)": [[6, "selenium_driverless.types.target.Target.find_element"]], "find_elements() (selenium_driverless.types.target.target method)": [[6, "selenium_driverless.types.target.Target.find_elements"]], "focus() (selenium_driverless.types.target.target method)": [[6, "selenium_driverless.types.target.Target.focus"]], "forward() (selenium_driverless.types.target.target method)": [[6, "selenium_driverless.types.target.Target.forward"]], "get() (selenium_driverless.types.target.target method)": [[6, "selenium_driverless.types.target.Target.get"]], "get_cdp_event_iter() (selenium_driverless.types.base_target.basetarget method)": [[6, "selenium_driverless.types.base_target.BaseTarget.get_cdp_event_iter"]], "get_cdp_event_iter() (selenium_driverless.types.target.target method)": [[6, "selenium_driverless.types.target.Target.get_cdp_event_iter"]], "get_cookie() (selenium_driverless.types.target.target method)": [[6, "selenium_driverless.types.target.Target.get_cookie"]], "get_cookies() (selenium_driverless.types.target.target method)": [[6, "selenium_driverless.types.target.Target.get_cookies"]], "get_history() (selenium_driverless.types.target.target method)": [[6, "selenium_driverless.types.target.Target.get_history"]], "get_issue_message() (selenium_driverless.types.target.target method)": [[6, "selenium_driverless.types.target.Target.get_issue_message"]], "get_network_conditions() (selenium_driverless.types.target.target method)": [[6, "selenium_driverless.types.target.Target.get_network_conditions"]], "get_screenshot_as_file() (selenium_driverless.types.target.target method)": [[6, "selenium_driverless.types.target.Target.get_screenshot_as_file"]], "get_screenshot_as_png() (selenium_driverless.types.target.target method)": [[6, "selenium_driverless.types.target.Target.get_screenshot_as_png"]], "get_sinks() (selenium_driverless.types.target.target method)": [[6, "selenium_driverless.types.target.Target.get_sinks"]], "get_target_for_iframe() (selenium_driverless.types.target.target method)": [[6, "selenium_driverless.types.target.Target.get_target_for_iframe"]], "get_targets_for_iframes() (selenium_driverless.types.target.target method)": [[6, "selenium_driverless.types.target.Target.get_targets_for_iframes"]], "id (selenium_driverless.types.target.targetinfo property)": [[6, "selenium_driverless.types.target.TargetInfo.id"]], "opener_frame_id (selenium_driverless.types.target.targetinfo property)": [[6, "selenium_driverless.types.target.TargetInfo.opener_frame_id"]], "opener_id (selenium_driverless.types.target.targetinfo property)": [[6, "selenium_driverless.types.target.TargetInfo.opener_id"]], "page_source (selenium_driverless.types.target.target property)": [[6, "selenium_driverless.types.target.Target.page_source"]], "pointer (selenium_driverless.types.target.target property)": [[6, "selenium_driverless.types.target.Target.pointer"]], "print_page() (selenium_driverless.types.target.target method)": [[6, "selenium_driverless.types.target.Target.print_page"]], "refresh() (selenium_driverless.types.target.target method)": [[6, "selenium_driverless.types.target.Target.refresh"]], "remove_cdp_listener() (selenium_driverless.types.base_target.basetarget method)": [[6, "selenium_driverless.types.base_target.BaseTarget.remove_cdp_listener"]], "remove_cdp_listener() (selenium_driverless.types.target.target method)": [[6, "selenium_driverless.types.target.Target.remove_cdp_listener"]], "save_screenshot() (selenium_driverless.types.target.target method)": [[6, "selenium_driverless.types.target.Target.save_screenshot"]], "save_snapshot() (selenium_driverless.types.target.target method)": [[6, "selenium_driverless.types.target.Target.save_snapshot"]], "search_elements() (selenium_driverless.types.target.target method)": [[6, "selenium_driverless.types.target.Target.search_elements"]], "send_keys() (selenium_driverless.types.target.target method)": [[6, "selenium_driverless.types.target.Target.send_keys"]], "set_network_conditions() (selenium_driverless.types.target.target method)": [[6, "selenium_driverless.types.target.Target.set_network_conditions"]], "set_sink_to_use() (selenium_driverless.types.target.target method)": [[6, "selenium_driverless.types.target.Target.set_sink_to_use"]], "set_source() (selenium_driverless.types.target.target method)": [[6, "selenium_driverless.types.target.Target.set_source"]], "snapshot() (selenium_driverless.types.target.target method)": [[6, "selenium_driverless.types.target.Target.snapshot"]], "socket (selenium_driverless.types.base_target.basetarget property)": [[6, "selenium_driverless.types.base_target.BaseTarget.socket"]], "socket (selenium_driverless.types.target.target property)": [[6, "selenium_driverless.types.target.Target.socket"]], "start_desktop_mirroring() (selenium_driverless.types.target.target method)": [[6, "selenium_driverless.types.target.Target.start_desktop_mirroring"]], "start_tab_mirroring() (selenium_driverless.types.target.target method)": [[6, "selenium_driverless.types.target.Target.start_tab_mirroring"]], "stop_casting() (selenium_driverless.types.target.target method)": [[6, "selenium_driverless.types.target.Target.stop_casting"]], "subtype (selenium_driverless.types.target.targetinfo property)": [[6, "selenium_driverless.types.target.TargetInfo.subtype"]], "title (selenium_driverless.types.target.target property)": [[6, "selenium_driverless.types.target.Target.title"]], "unfocus() (selenium_driverless.types.target.target method)": [[6, "selenium_driverless.types.target.Target.unfocus"]], "wait_download() (selenium_driverless.types.target.target method)": [[6, "selenium_driverless.types.target.Target.wait_download"]], "wait_for_cdp() (selenium_driverless.types.base_target.basetarget method)": [[6, "selenium_driverless.types.base_target.BaseTarget.wait_for_cdp"]], "wait_for_cdp() (selenium_driverless.types.target.target method)": [[6, "selenium_driverless.types.target.Target.wait_for_cdp"]], "xhr() (selenium_driverless.types.target.target method)": [[6, "selenium_driverless.types.target.Target.xhr"]]}})
+Search.setIndex({"docnames": ["api/By", "api/Chrome", "api/ChromeOptions", "api/Context", "api/Input", "api/RequestInterception", "api/Target", "api/WebELement", "index"], "filenames": ["api\\By.rst", "api\\Chrome.rst", "api\\ChromeOptions.rst", "api\\Context.rst", "api\\Input.rst", "api\\RequestInterception.rst", "api\\Target.rst", "api\\WebELement.rst", "index.rst"], "titles": ["By Element Locator", "webdriver.Chrome", "ChromeOptions", "Context", "Input", "Request-Interception", "Target", "WebElement", "Documentation of Driverless"], "terms": {"class": [0, 1, 2, 3, 4, 5, 6, 7], "selenium_driverless": [0, 1, 2, 3, 4, 5, 6, 7, 8], "type": [0, 1, 2, 3, 4, 5, 6, 7, 8], "sourc": [0, 1, 2, 3, 4, 5, 6, 7], "set": [0, 1, 2, 3, 5, 6, 7], "support": [0, 1, 2, 3, 6], "strategi": 0, "id": [0, 1, 5, 6, 7], "name": [0, 1, 2, 3, 5, 6, 7], "xpath": [0, 3, 6, 8], "tag_nam": [0, 3, 7], "tag": 0, "class_nam": [0, 7], "css_selector": 0, "css": 0, "selector": [0, 3, 6], "alia": [0, 1, 3, 6, 7], "option": [1, 2, 3, 4, 5, 6, 7, 8], "none": [1, 2, 3, 4, 5, 6, 7], "timeout": [1, 2, 3, 4, 5, 6, 7, 8], "30": [1, 3, 5, 6, 7], "debug": 1, "fals": [1, 2, 3, 5, 6, 7], "max_ws_siz": [1, 3, 5, 6], "134217728": 1, "control": 1, "chromium": [1, 2, 3, 6], "base": [1, 3, 6], "browser": [1, 2, 3, 5, 6], "without": [1, 3, 4, 5], "ani": [1, 3, 5, 6, 7], "driver": [1, 2, 3, 4, 5, 6, 8], "properti": [1, 2, 3, 5, 6, 7], "frame_tre": 1, "dict": [1, 2, 3, 4, 5, 6, 7], "async": [1, 2, 3, 4, 5, 6, 7, 8], "all": [1, 2, 3, 6, 7, 8], "nest": [1, 7], "frame": [1, 3, 5, 6, 7], "within": [1, 3, 6], "current": [1, 3, 6, 7], "target": [1, 3, 4, 5, 7, 8], "str": [1, 2, 3, 4, 5, 6, 7], "targetinfo": [1, 3, 6, 8], "context": [1, 6, 8], "incognito": [1, 6], "new_context": 1, "proxy_bypass_list": 1, "proxy_serv": 1, "true": [1, 2, 3, 4, 5, 6, 7, 8], "universal_access_origin": 1, "url": [1, 2, 3, 5, 6, 7], "about": [1, 2], "blank": [1, 2], "creat": [1, 2, 3, 5], "new": [1, 3, 5], "paramet": [1, 2, 3, 4, 5, 6, 7], "first": [1, 2, 7], "tab": [1, 2, 3, 6], "start": [1, 2, 3, 5, 6], "default": [1, 2, 3, 4, 5, 6], "list": [1, 2, 3, 4, 5, 6, 7], "an": [1, 2, 3, 4, 5, 6, 7], "origin": [1, 3, 5, 6], "grant": 1, "unlimit": 1, "cross": [1, 5], "access": [1, 6], "part": 1, "other": [1, 3, 5], "than": 1, "those": [1, 6], "constitut": 1, "ar": [1, 6], "ignor": [1, 2], "return": [1, 2, 3, 4, 5, 6, 7], "The": [1, 2, 3, 5, 6, 7], "proxi": [1, 2, 5], "doesn": [1, 2, 3, 5, 7], "t": [1, 2, 3, 5, 6, 7, 8], "work": 1, "window": [1, 3, 6, 7], "due": [1, 2, 4], "crbug": [1, 2], "1310057": 1, "server": [1, 2, 5], "us": [1, 2, 3, 4, 5, 6, 7], "get_target": 1, "_type": 1, "context_id": [1, 3, 6, 7], "self": 1, "get": [1, 2, 3, 5, 6, 7, 8], "liter": [1, 3, 5, 6, 7], "page": [1, 3, 5, 6, 7, 8], "background_pag": 1, "service_work": 1, "param": [1, 5, 6], "filter": 1, "thi": [1, 2, 3, 4, 5, 6, 7, 8], "function": [1, 3, 4, 6], "current_target": 1, "base_target": [1, 3, 6], "basetarget": [1, 5, 6, 8], "connect": [1, 2, 6, 7], "handl": [1, 3, 5, 7], "global": [1, 6], "onli": [1, 2, 3, 6, 7], "bind": 1, "cdp": [1, 3, 4, 6], "protocol": [1, 3, 6], "mv3_extens": [1, 2], "background": [1, 3], "script": [1, 3, 6, 7, 8], "load": [1, 2, 3, 5, 6], "extens": [1, 2, 3, 5, 6], "manifest": 1, "version": 1, "3": [1, 3, 6, 7], "ensure_extensions_incognito_allow": 1, "ensur": [1, 7], "instal": 1, "allow": [1, 3, 6], "gener": [1, 7], "decid": 1, "whether": [1, 2, 3, 4, 5, 6, 7], "split": 1, "span": 1, "not_allow": 1, "configur": 1, "For": [1, 3, 6], "chang": [1, 5, 7], "behaviour": [1, 3], "you": [1, 2, 3, 5, 6, 7], "ll": 1, "have": [1, 6], "modifi": [1, 4, 8], "json": [1, 2, 3, 4, 5, 6, 7], "file": [1, 2, 3, 6, 7], "compress": [1, 2, 5], "directori": [1, 2, 3, 6], "see": [1, 3, 5, 6, 7, 8], "develop": [1, 6], "com": [1, 2, 8], "doc": [1, 6], "refer": [1, 3, 5, 6, 7], "base_context": 1, "which": [1, 2, 4, 5, 6], "isn": [1, 7], "downloads_dir": [1, 2, 3, 6], "download": [1, 2, 3, 6], "set_download_behaviour": [1, 3], "path": [1, 2, 3, 4, 6, 7], "deni": [1, 3], "allowandnam": [1, 3], "instead": [1, 2, 3, 6, 7], "can": [1, 2, 3, 4, 6, 7], "caus": [1, 3, 5], "some": [1, 3, 5, 8], "bug": [1, 3], "current_context": 1, "switch": [1, 3], "target_id": [1, 3, 6], "2": [1, 3, 4, 5, 6, 7, 8], "targetid": [1, 6], "advanc": 1, "usag": [1, 3, 6], "float": [1, 3, 4, 6, 7], "second": [1, 3, 4, 6, 7], "": [1, 7], "track": 1, "alreadi": [1, 5], "get_target_for_ifram": [1, 6], "ifram": [1, 3, 6, 7], "specif": [1, 3, 4, 6], "oopif": 1, "site": [1, 5], "isol": [1, 6], "solut": 1, "look": 1, "webel": [1, 3, 4, 6, 8], "content_docu": [1, 6, 7], "get_targets_for_ifram": [1, 6], "more": 1, "inform": 1, "wait_download": [1, 3, 6], "wait": [1, 3, 6, 7, 8], "someth": [1, 3, 6], "like": [1, 3, 6, 7], "frameid": [1, 3, 5, 6], "2d543b5e8b14945b280c537a4882a695": [1, 3, 6], "guid": [1, 3, 6], "c91df4d5": [1, 3, 6], "9b45": [1, 3, 6], "4962": [1, 3, 6], "84df": [1, 3, 6], "3749bd3f926d": [1, 3, 6], "http": [1, 2, 3, 5, 6, 7, 8], "www": [1, 2, 3, 6], "w3": [1, 2, 3, 5, 6], "org": [1, 2, 3, 5, 6, 7], "wai": [1, 2, 3, 6], "er": [1, 2, 3, 6], "test": [1, 2, 3, 6, 7], "xhtml": [1, 2, 3, 6], "testfil": [1, 2, 3, 6], "resourc": [1, 2, 3, 5, 6], "pdf": [1, 2, 3, 6], "dummi": [1, 2, 3, 6], "suggestedfilenam": [1, 3, 6], "specifi": [1, 3, 4, 5, 6], "guid_fil": [1, 2, 3, 6], "d": [1, 3, 6], "system": [1, 3, 6], "appdata": [1, 3, 6], "pycharm": [1, 3, 6], "scratch": [1, 3, 6], "time": [1, 3, 4, 5, 6, 7], "from": [1, 2, 3, 4, 5, 6, 7, 8], "yet": [1, 2, 3, 6, 7, 8], "referr": [1, 3, 5, 6], "wait_load": [1, 3, 5, 6, 8], "web": [1, 3, 6, 7], "bool": [1, 2, 3, 4, 5, 6, 7], "webpag": [1, 3, 6, 7], "maximum": [1, 3, 6, 7], "same": [1, 3, 5, 6], "initi": [1, 3, 5, 6], "titl": [1, 3, 6, 8], "current_point": [1, 3], "pointer": [1, 3, 4, 6, 8], "send_kei": [1, 3, 6, 7], "text": [1, 3, 4, 5, 6, 7, 8], "send": [1, 3, 6, 7], "kei": [1, 3, 5, 6, 7], "execute_raw_script": [1, 3, 6, 7], "arg": [1, 3, 6, 7], "await_r": [1, 3, 6], "serial": [1, 3, 6, 7], "deep": [1, 6], "max_depth": [1, 3, 6, 7], "execution_context_id": [1, 3, 6, 7], "unique_context": [1, 3, 6, 7], "execut": [1, 2, 3, 6, 7], "javascript": [1, 3, 6, 7], "globalthi": [1, 3, 6, 7], "argument": [1, 2, 3, 6, 7], "document": [1, 3, 6, 7], "obj": [1, 3, 6, 7], "here": [1, 3, 6, 7], "string": [1, 5, 6], "pass": [1, 2, 5, 6, 7], "either": [1, 2, 4, 6], "serializ": [1, 6], "remoteobject": [1, 6], "await": [1, 2, 3, 4, 5, 6, 7, 8], "valu": [1, 2, 3, 4, 5, 6, 7], "idonli": [1, 6], "one": [1, 3, 5, 6, 7], "int": [1, 2, 3, 4, 5, 6, 7], "depth": [1, 6, 7], "object": [1, 3, 5, 6], "complet": [1, 6, 8], "run": [1, 2, 5, 6, 8], "exclus": [1, 6], "runtim": [1, 6, 7], "callfunctionon": [1, 6], "execute_script": [1, 3, 6, 7], "synchron": [1, 3, 6, 7], "descript": [1, 3, 6, 7], "execute_async_script": [1, 3, 6, 7], "asynchron": [1, 3, 6, 7], "i": [1, 2, 3, 4, 5, 6, 7, 8], "recommend": [1, 3, 4, 7], "except": [1, 3, 7], "correctli": [1, 3, 7], "eval_async": [1, 3, 6, 7], "resolv": [1, 3, 6, 7], "length": [1, 3, 6, 7], "1": [1, 3, 4, 5, 6, 7], "re": [1, 3, 6, 7], "fetch": [1, 3, 5, 6, 7], "httpbin": [1, 3, 5, 6, 7], "mind": [1, 3, 5, 6, 7], "cor": [1, 3, 6, 7], "current_url": [1, 3, 6], "page_sourc": [1, 3, 6], "html": [1, 6, 7, 8], "close": [1, 3, 6, 7], "focu": [1, 3, 6, 7], "focus": [1, 7], "quit": [1, 3], "clean_dir": 1, "clean": [1, 2], "out": 1, "user": [1, 2, 5, 6, 7], "data": [1, 2, 3, 5, 6, 7], "dir": [1, 2], "current_target_info": 1, "current_window_handl": [1, 3], "deprec": [1, 2], "remov": [1, 2, 3, 6, 7], "current_window_id": 1, "windowid": 1, "belong": 1, "window_handl": [1, 3], "aren": [1, 6, 8], "order": [1, 5], "posit": [1, 3, 4], "do": [1, 7], "reli": 1, "index": 1, "iter": [1, 3, 5, 6], "them": 1, "new_window": [1, 3], "type_hint": [1, 3], "activ": [1, 3, 6], "should": [1, 2, 3, 4, 5, 6], "explicitli": 1, "new_target": 1, "set_window_st": 1, "state": 1, "normal": [1, 4], "minim": [1, 3], "maxim": [1, 3, 6], "fullscreen": 1, "normalize_window": 1, "size": [1, 7], "maximize_window": [1, 3], "fullscreen_window": [1, 3], "enter": [1, 5], "minimize_window": [1, 3], "rtype": [1, 2], "throttl": 1, "print_pag": [1, 3, 6], "print": [1, 2, 5, 6, 8], "switch_to": [1, 3, 8], "switchto": [1, 3, 8], "back": [1, 3, 4, 6], "goe": [1, 3, 6], "step": [1, 3, 6], "backward": [1, 3, 6], "histori": [1, 3, 6], "ha": [1, 2, 5, 6, 7], "forward": [1, 3, 4, 6], "refresh": [1, 3, 6], "get_cooki": [1, 3, 6], "cooki": [1, 3, 6], "singl": [1, 2, 3, 5, 6], "found": [1, 3, 6], "delete_cooki": [1, 3, 6], "domain": [1, 3, 6], "delet": [1, 3, 6], "given": [1, 3, 6, 7], "delete_all_cooki": [1, 3, 6], "add_cooki": [1, 3, 6], "cookie_dict": [1, 3, 6], "add": [1, 2, 3, 6], "network": [1, 3, 5, 6], "cookieparam": [1, 6], "static": [1, 5], "sleep": [1, 4, 8], "time_to_wait": 1, "sync": 1, "find_el": [1, 3, 6, 7, 8], "find": [1, 3, 6, 7], "element": [1, 3, 6, 7, 8], "locat": [1, 3, 4, 6, 7, 8], "By": [1, 2, 3, 5, 6, 7, 8], "actual": [1, 3, 6, 7], "queri": [1, 3, 6, 7], "how": [1, 3, 4, 6, 7], "long": [1, 3, 5, 6, 7], "exist": [1, 3, 6, 7, 8], "multipl": [1, 2, 3, 6, 7], "now": 6, "being": [1, 2, 3, 6, 7], "reload": [1, 3, 6], "loop": [1, 3, 6, 7], "search_el": [1, 3, 6], "similarli": 1, "ctrl": [1, 4], "f": 1, "devtool": [1, 2, 3, 6], "consol": 1, "plain": [1, 3, 6], "get_screenshot_as_fil": [1, 3, 6], "filenam": [1, 3, 6, 7], "save": [1, 2, 3, 6, 7], "screenshot": [1, 3, 6, 7], "png": [1, 3, 6, 7], "imag": [1, 3, 6, 7], "wish": [1, 3, 6], "your": [1, 3, 5, 6], "end": [1, 3, 6], "save_screenshot": [1, 3, 6], "func": [1, 3, 6], "get_screenshot_as_png": [1, 3, 6], "binari": [1, 2, 3, 5, 6, 7], "byte": [1, 3, 5, 6, 7], "snapshot": [1, 3, 6], "mhtml": [1, 3, 6], "save_snapshot": [1, 3, 6], "full": [1, 3, 6], "get_snapshot": [1, 3, 6], "set_window_s": [1, 3], "width": [1, 3], "height": [1, 3], "unless": [1, 2], "windowhandl": [1, 3], "pixel": [1, 3], "get_window_s": [1, 3], "code": [1, 5], "block": [1, 5], "1280": 1, "720": 1, "set_window_posit": [1, 3], "x": [1, 3, 4], "y": [1, 3, 4], "coordin": [1, 3, 4, 7], "get_window_posit": [1, 3], "0": [1, 3, 4, 5, 7, 8], "get_window_rect": [1, 3], "windowst": 1, "set_window_rect": [1, 3], "get_network_condit": [1, 3, 6], "emul": [1, 3, 6], "latenc": [1, 3, 6], "4": [1, 3, 4, 6], "download_throughput": [1, 3, 6], "upload_throughput": [1, 3, 6], "offlin": [1, 3, 6], "set_network_condit": [1, 3, 6], "connection_typ": [1, 3, 6], "addit": [1, 3, 6], "m": [1, 3, 6, 8], "throughput": [1, 3, 6], "500": [1, 3, 6], "1024": [1, 3, 6], "exampl": [1, 2, 3, 6, 8], "cellular2g": 1, "cellular3g": 1, "cellular4g": 1, "bluetooth": 1, "ethernet": 1, "wifi": [1, 3, 6], "wimax": 1, "both": [1, 3, 6], "upload": [1, 3, 6], "delete_network_condit": [1, 3, 6], "reset": [1, 3, 6], "set_permiss": [1, 3], "applic": [1, 3, 5], "permiss": [1, 3], "item": [1, 3], "prompt": 1, "appli": [1, 2], "clipboard": [1, 3], "read": [1, 3], "set_proxi": 1, "proxy_config": 1, "dynam": [1, 6], "mode": [1, 6], "fixed_serv": 1, "rule": 1, "proxyforhttp": 1, "scheme": [1, 5], "host": [1, 2, 6], "port": [1, 2], "proxyforftp": 1, "fallbackproxi": 1, "bypasslist": 1, "local": 1, "authentif": [1, 5], "set_auth": 1, "set_single_proxi": 1, "bypass_list": 1, "user1": [1, 2], "passwrd1": [1, 2], "5001": [1, 2], "when": [1, 2, 3, 5, 6], "been": [1, 2, 5], "driverless": [1, 2], "socks5": [1, 2], "authent": [1, 2, 5, 6], "1309413": [1, 2], "clear_proxi": 1, "clear": [1, 7], "usernam": [1, 5], "password": [1, 5, 6], "host_with_port": 1, "format": [1, 2, 7], "clear_auth": 1, "auth": [1, 2, 5], "wait_for_cdp": [1, 3, 6, 8], "event": [1, 3, 4, 5, 6, 7], "add_cdp_listen": [1, 3, 6], "callback": [1, 3, 5, 6], "listen": [1, 3, 6, 7], "remove_cdp_listen": [1, 3, 6], "get_cdp_event_it": [1, 3, 6], "over": [1, 3, 5, 6, 7], "asynciter": [1, 3, 5, 6], "execute_cdp_cmd": [1, 3, 6], "cmd": [1, 3, 6], "cmd_arg": [1, 3, 6], "10": [1, 3, 6, 8], "command": [1, 3, 4, 6, 7], "messag": [1, 3, 6], "Not": [1, 2, 3, 7], "receiv": [1, 3, 6], "kwarg": [1, 3, 4], "j": [1, 3, 6], "request": [1, 2, 3, 6, 8], "xhr": [1, 3, 6], "xmlhttprequest": [1, 3, 6], "get_sink": [1, 3, 6], "A": [1, 3, 6, 7], "sink": [1, 3, 6], "avail": [1, 3, 6, 7], "cast": [1, 3, 6], "get_issue_messag": [1, 3, 6], "error": [1, 3, 5, 6], "issu": [1, 3, 6], "session": [1, 2, 3, 6], "set_sink_to_us": [1, 3, 6], "sink_nam": [1, 3, 6], "its": [1, 3, 6], "start_desktop_mirror": [1, 3, 6], "desktop": [1, 3, 6], "mirror": [1, 3, 6], "start_tab_mirror": [1, 3, 6], "stop_cast": [1, 3, 6], "stop": [1, 3, 6], "alert": [1, 3, 8], "get_alert": 1, "5": [1, 2, 3, 4, 6, 7, 8], "default_cont": [1, 3], "frame_refer": 1, "union": [1, 2, 4, 5, 7], "bring": [1, 3, 6], "front": [1, 3, 6], "window_id": 1, "what": [1, 3], "kind": [1, 3], "webdriv": [2, 3, 5, 6, 8], "reus": 2, "use_extens": 2, "don": 2, "chrome": [2, 3, 5, 6, 7, 8], "while": 2, "requir": [2, 3, 4, 5], "ad": 2, "As": 2, "altern": 2, "might": [2, 5, 6, 7], "intercept": [2, 8], "provid": [2, 3, 5, 6], "add_argu": 2, "launch": 2, "pref": 2, "prefer": 2, "update_pref": 2, "updat": 2, "dot": 2, "user_data_dir": 2, "temporarili": 2, "temp": 2, "_dir": 2, "o": 2, "getcwd": 2, "isdir": 2, "mkdir": 2, "plugin": 2, "always_open_pdf_extern": 2, "download_data": 2, "absolut": [2, 7], "headless": 2, "startup_url": 2, "binary_loc": 2, "env": 2, "subprocess": 2, "popen": 2, "environ": 2, "add_extens": 2, "zip": 2, "crx": 2, "etc": [2, 6], "extract": 2, "debugger_address": 2, "address": 2, "remot": 2, "instanc": [2, 4, 6, 7], "make": [2, 3, 5, 6], "well": [2, 3], "single_proxi": 2, "auto_clean_dir": 2, "automat": [2, 5, 7], "enable_mobil": 2, "android_packag": 2, "android": 2, "android_act": 2, "device_seri": 2, "enabl": [2, 7], "mobil": 2, "packag": 2, "implement": [2, 7], "accept_insecure_cert": 2, "accept": [2, 6, 7, 8], "insecur": 2, "certif": 2, "notimpl": [2, 7], "ignore_local_proxy_environment_vari": 2, "call": [2, 7], "http_proxi": 2, "https_proxi": 2, "pick": 2, "up": [2, 4], "add_experimental_opt": 2, "experiment": 2, "method": [2, 3, 5, 6, 7, 8], "is_incognito": 3, "1048576": [3, 6], "drive": 3, "chromedriv": 3, "behavior": [3, 5], "click": [3, 4, 7, 8], "const": 3, "accord": 3, "obj_id": [3, 7], "docs_sourc": 3, "start_monoton": 3, "everi": [3, 7], "associ": 3, "invok": 3, "manag": 3, "screen": 3, "oper": [3, 7], "take": [3, 4, 5, 6, 7], "best": [3, 6], "effort": [3, 6], "contain": 3, "active_el": 3, "frame_nam": 3, "parent_fram": 3, "main": [3, 4, 5, 8], "dictionari": [3, 6, 7], "correspond": [3, 5, 6], "visibl": [3, 6, 7], "my_cooki": 3, "scope": [3, 6], "secur": 3, "httponli": 3, "expiri": 3, "samesit": 3, "foo": 3, "bar": 3, "strict": [3, 5], "search": [3, 6], "resizeto": 3, "800": 3, "600": 3, "moveto": 3, "w3c": 3, "compat": 3, "100": 3, "200": [3, 5], "network_condit": [3, 6], "condit": [3, 6], "note": [3, 5, 6], "pointer_typ": 4, "mous": [4, 7], "down": 4, "press": 4, "pointerev": [4, 8], "releas": 4, "x_or_elem": 4, "move_to": [4, 7, 8], "move_kwarg": 4, "click_kwarg": [4, 7], "move": [4, 7], "basepoint": [4, 8], "total_tim": 4, "accel": 4, "mid_tim": 4, "smooth_soft": 4, "20": [4, 6], "pinter": 4, "total": 4, "acceler": 4, "deceler": 4, "pointermov": 4, "perform": [4, 7], "where": [4, 5], "half": 4, "curvi": 4, "line": 4, "move_path": 4, "dispatch": 4, "between": 4, "mouseup": [4, 7], "mousedown": [4, 7], "double_click": 4, "doubl": 4, "pos_from_time_callback": 4, "freq_assumpt": 4, "60": 4, "assumpt": 4, "mousemov": 4, "frequenc": 4, "accuraci": 4, "callabl": [4, 5, 6], "tupl": 4, "type_": 4, "timestamp": 4, "button": [4, 7, 8], "left": 4, "click_count": 4, "forc": 4, "tangential_pressur": 4, "tilt_x": 4, "tilt_i": 4, "twist": 4, "delta_x": 4, "delta_i": 4, "to_json": 4, "alt": 4, "shift": 4, "8": 4, "pointertyp": [4, 8], "mousepoint": 4, "pen": 4, "mousebutton": [4, 8], "middl": [4, 7], "right": 4, "16": 4, "eventtyp": [4, 8], "mousepress": 4, "mousereleas": 4, "wheel": 4, "mousewheel": 4, "import": [5, 8], "asyncio": [5, 8], "network_interceptor": 5, "networkinterceptor": 5, "interceptedrequest": 5, "interceptedauth": 5, "requestpattern": 5, "requeststag": 5, "def": [5, 8], "on_request": 5, "post": [5, 6], "continue_request": 5, "basic": 5, "intercept_respons": 5, "pattern": 5, "anyrequest": 5, "intercept_auth": 5, "interceptor": 5, "ensure_futur": 5, "isinst": 5, "virtual": 5, "zero": 5, "would": 5, "continue_auth": 5, "elif": 5, "stage": 5, "respons": [5, 6], "bodi": [5, 6, 8], "break": 5, "on_respons": 5, "on_auth": 5, "bypass_service_work": 5, "__init__": 5, "onrequest": 5, "onrespons": 5, "onauth": 5, "bypass": 5, "servic": 5, "worker": 5, "__aiter__": 5, "els": [5, 7], "possibl": [5, 7], "patter": 5, "is_redirect": 5, "redirect": [5, 6], "response_status_cod": 5, "301": 5, "302": 5, "303": 5, "307": 5, "308": 5, "requestpaus": 5, "frame_id": [5, 7], "requestid": [5, 6], "resource_typ": 5, "resourcetyp": 5, "network_id": 5, "requestwillbes": 5, "had": 5, "fire": 5, "response_error_reason": 5, "reason": 5, "response_head": 5, "header": [5, 6], "statu": [5, 6], "response_status_text": 5, "redirected_id": 5, "bypass_brows": 5, "allow_redirect": 5, "proxy_auth": 5, "extern": 5, "doe": 5, "tl": 5, "fingerprint": 5, "accordingli": 5, "technic": 5, "detect": [4, 5, 7], "post_data": 5, "continu": 5, "fulfil": 5, "arrai": 5, "observ": 5, "overrid": 5, "continue_respons": 5, "binary_response_head": 5, "response_cod": 5, "response_phras": 5, "separ": 5, "seri": 5, "pair": 5, "treat": 5, "base64": [5, 6, 7], "encod": [5, 6], "phrase": 5, "ok": [5, 6], "resum": 5, "rais": 5, "cancel": 5, "done": 5, "fail_request": 5, "error_reason": 5, "fail": [5, 6, 7], "abort": 5, "timedout": 5, "accessdeni": 5, "connectionclos": 5, "connectionreset": 5, "connectionrefus": 5, "connectionabort": 5, "connectionfail": 5, "namenotresolv": 5, "internetdisconnect": 5, "addressunreach": 5, "blockedbycli": 5, "blockedbyrespons": 5, "authrequir": 5, "auth_challeng": 5, "authchalleng": 5, "cancelauth": 5, "providecredenti": 5, "challeng": 5, "digest": 5, "realm": 5, "pleas": 5, "fragment": 5, "url_frag": 5, "hash": 5, "present": 5, "has_post_data": 5, "postdata": 5, "still": [5, 7], "omit": 5, "flag": 5, "too": 5, "post_data_entri": 5, "postdataentri": 5, "mixed_content_typ": 5, "blockabl": 5, "mix": 5, "content": 5, "initial_prior": 5, "verylow": 5, "low": 5, "medium": 5, "high": [5, 6], "veryhigh": 5, "prioriti": [5, 6], "sent": 5, "referrer_polici": [5, 6], "unsaf": 5, "downgrad": 5, "polici": 5, "defin": 5, "tr": 5, "is_link_preload": 5, "via": 5, "link": [5, 6], "preload": 5, "trust_token_param": 5, "trusttokenparam": 5, "is_same_sit": 5, "consid": [5, 7], "enumer": 5, "urlpattern": 5, "anyrespons": 5, "url_pattern": 5, "request_stag": 5, "authalreadyhandledexcept": 5, "requestdoneexcept": 5, "is_remot": 6, "start_socket": 6, "usual": 6, "webwork": 6, "socket": 6, "singlecdpsocket": 6, "own": 6, "allow_not_on_map": 6, "keyboard": 6, "map": [6, 7], "unfocu": 6, "disabl": [6, 7], "get_histori": 6, "getnavigationhistori": 6, "navigationhistori": 6, "set_sourc": [6, 7], "15": [6, 8], "outerhtml": [6, 7], "dom": [6, 7], "try": 6, "framenavig": 6, "result": 6, "follow": 6, "chromedevtool": 6, "github": [6, 8], "io": 6, "empti": 6, "getresponsebodi": 6, "base64encod": 6, "credenti": 6, "cach": 6, "integr": 6, "keepal": 6, "mozilla": 6, "en": 6, "u": 6, "api": 6, "smth": 6, "status_cod": 6, "status_text": 6, "with_credenti": 6, "mime_typ": 6, "extra_head": 6, "put": 6, "includ": 6, "pars": 6, "extra": 6, "responsetext": 6, "responsetyp": 6, "responseurl": 6, "responsexml": 6, "statustext": 6, "responsehead": 6, "target_info": 6, "target_gett": 6, "info": 6, "itself": 6, "attach": [6, 7], "client": 6, "opener_id": 6, "open": 6, "can_access_open": 6, "opener_frame_id": 6, "browser_context_id": 6, "browsercontextid": 6, "subtyp": 6, "detail": [6, 7], "mai": 6, "portal": 6, "prerend": 6, "chromeinst": 6, "repres": [6, 7], "whole": 6, "unfortun": 6, "downloads_dir_for_context": 6, "isolated_exec_id": 7, "node_id": 7, "backend_node_id": 7, "is_ifram": 7, "interest": 7, "interact": 7, "through": 7, "interfac": 7, "fresh": 7, "check": 7, "valid": 7, "essenti": 7, "determin": 7, "If": 7, "staleelementreferenceexcept": 7, "thrown": 7, "futur": 7, "remoteobjectid": 7, "executioncontextid": 7, "nodeid": 7, "contentdocu": 7, "shadow_root": 7, "shadow": 7, "root": 7, "document_url": 7, "backendnodeid": 7, "classnam": 7, "idx": 7, "get_listen": 7, "domdebugg": 7, "geteventlisten": 7, "get_properti": 7, "elem": [4, 7, 8], "tagnam": 7, "textcont": 7, "entri": 7, "highlight": 7, "fade": 7, "is_click": 7, "listener_depth": 7, "input": [7, 8], "select": [7, 8], "textarea": 7, "video": 7, "wise": 7, "visible_timeout": 7, "spread_a": 7, "spread_b": 7, "bias_a": 7, "bias_b": 7, "border": 7, "05": 7, "scroll_to": 7, "ensure_click": 7, "abl": 7, "comput": 7, "box": 7, "model": 7, "spread": 7, "b": 7, "bia": 7, "minimum": 7, "toward": 7, "edg": 7, "rel": 7, "random": 7, "point": 7, "outsid": 7, "scroll": 7, "clickabl": 7, "reliabl": 7, "equival": 7, "6": 7, "std": 7, "99": 7, "7": 7, "write": 7, "click_on": 7, "insert": 7, "befor": 7, "set_fil": 7, "mid_loc": 7, "probabl": 7, "submit": 7, "form": 7, "It": 7, "dom_attribut": 7, "attribut": 7, "properli": [7, 8], "welel": 7, "get_dom_attribut": 7, "declar": 7, "markup": 7, "retriev": 7, "set_dom_attribut": 7, "get_attribut": 7, "is_select": 7, "checkbox": 7, "radio": 7, "is_en": 7, "is_displai": 7, "location_once_scrolled_into_view": 7, "rect": 7, "value_of_css_properti": 7, "property_nam": 7, "render": 7, "canva": 7, "box_model": 7, "boxmodel": 7, "aria_rol": 7, "aria": 7, "role": 7, "accessible_nam": 7, "level": 7, "screenshot_as_base64": 7, "screenshot_as_png": 7, "parent": 7, "python": 8, "pip": 8, "upgrad": 8, "selenium": 8, "chromeopt": 8, "nowsecur": 8, "nl": 8, "relax": 8, "domcontenteventfir": 8, "div": 8, "p": 8, "kaliiiiiiiiii": 8, "util": 4, "async_input": 4, "01": 4, "differ": 4, "asyncinput": 4, "cdp_patch": 4, "action": 4, "potenti": 4, "patch": 4}, "objects": {"selenium_driverless.input.pointer": [[4, 0, 1, "", "BasePointer"], [4, 0, 1, "", "Buttons"], [4, 0, 1, "", "EventType"], [4, 0, 1, "", "Modifiers"], [4, 0, 1, "", "MouseButton"], [4, 0, 1, "", "Pointer"], [4, 0, 1, "", "PointerEvent"], [4, 0, 1, "", "PointerType"]], "selenium_driverless.input.pointer.BasePointer": [[4, 1, 1, "", "click"], [4, 1, 1, "", "dispatch"], [4, 1, 1, "", "double_click"], [4, 1, 1, "", "down"], [4, 1, 1, "", "move_path"], [4, 1, 1, "", "move_to"], [4, 1, 1, "", "up"]], "selenium_driverless.input.pointer.Buttons": [[4, 2, 1, "", "BACK"], [4, 2, 1, "", "DEFAULT"], [4, 2, 1, "", "FORWARD"], [4, 2, 1, "", "LEFT"], [4, 2, 1, "", "MIDDLE"], [4, 2, 1, "", "NONE"], [4, 2, 1, "", "RIGHT"]], "selenium_driverless.input.pointer.EventType": [[4, 2, 1, "", "MOVE"], [4, 2, 1, "", "PRESS"], [4, 2, 1, "", "RELEASE"], [4, 2, 1, "", "WHEEL"]], "selenium_driverless.input.pointer.Modifiers": [[4, 2, 1, "", "ALT"], [4, 2, 1, "", "COMMAND"], [4, 2, 1, "", "CTRL"], [4, 2, 1, "", "NONE"], [4, 2, 1, "", "SHIFT"]], "selenium_driverless.input.pointer.MouseButton": [[4, 2, 1, "", "BACK"], [4, 2, 1, "", "FORWARD"], [4, 2, 1, "", "LEFT"], [4, 2, 1, "", "MIDDLE"], [4, 2, 1, "", "NONE"], [4, 2, 1, "", "RIGHT"]], "selenium_driverless.input.pointer.Pointer": [[4, 1, 1, "", "click"], [4, 1, 1, "", "down"], [4, 1, 1, "", "move_to"], [4, 1, 1, "", "up"]], "selenium_driverless.input.pointer.PointerEvent": [[4, 1, 1, "", "to_json"]], "selenium_driverless.input.pointer.PointerType": [[4, 2, 1, "", "MOUSE"], [4, 2, 1, "", "PEN"]], "selenium_driverless.input.utils": [[4, 3, 1, "", "select"]], "selenium_driverless.scripts.network_interceptor": [[5, 0, 1, "", "AuthAlreadyHandledException"], [5, 0, 1, "", "AuthChallenge"], [5, 0, 1, "", "InterceptedAuth"], [5, 0, 1, "", "InterceptedRequest"], [5, 0, 1, "", "NetworkInterceptor"], [5, 0, 1, "", "Request"], [5, 0, 1, "", "RequestDoneException"], [5, 0, 1, "", "RequestPattern"], [5, 0, 1, "", "RequestStages"]], "selenium_driverless.scripts.network_interceptor.AuthAlreadyHandledException": [[5, 4, 1, "", "request"]], "selenium_driverless.scripts.network_interceptor.AuthChallenge": [[5, 4, 1, "", "origin"], [5, 4, 1, "", "params"], [5, 4, 1, "", "realm"], [5, 4, 1, "", "scheme"], [5, 4, 1, "", "source"], [5, 4, 1, "", "target"]], "selenium_driverless.scripts.network_interceptor.InterceptedAuth": [[5, 4, 1, "", "auth_challenge"], [5, 1, 1, "", "cancel"], [5, 1, 1, "", "continue_auth"], [5, 4, 1, "", "frame_id"], [5, 4, 1, "", "id"], [5, 4, 1, "", "params"], [5, 4, 1, "", "request"], [5, 4, 1, "", "resource_type"], [5, 1, 1, "", "resume"], [5, 4, 1, "", "target"]], "selenium_driverless.scripts.network_interceptor.InterceptedRequest": [[5, 4, 1, "", "body"], [5, 1, 1, "", "bypass_browser"], [5, 1, 1, "", "continue_request"], [5, 1, 1, "", "continue_response"], [5, 1, 1, "", "fail_request"], [5, 4, 1, "", "frame_id"], [5, 1, 1, "", "fulfill"], [5, 4, 1, "", "id"], [5, 4, 1, "", "is_redirect"], [5, 4, 1, "", "network_id"], [5, 4, 1, "", "params"], [5, 4, 1, "", "redirected_id"], [5, 4, 1, "", "request"], [5, 4, 1, "", "resource_type"], [5, 4, 1, "", "response_error_reason"], [5, 4, 1, "", "response_headers"], [5, 4, 1, "", "response_status_code"], [5, 4, 1, "", "response_status_text"], [5, 1, 1, "", "resume"], [5, 4, 1, "", "stage"], [5, 4, 1, "", "target"]], "selenium_driverless.scripts.network_interceptor.NetworkInterceptor": [[5, 1, 1, "", "__aiter__"], [5, 1, 1, "", "__init__"], [5, 4, 1, "", "patterns"], [5, 4, 1, "", "target"]], "selenium_driverless.scripts.network_interceptor.Request": [[5, 4, 1, "", "has_post_data"], [5, 4, 1, "", "headers"], [5, 4, 1, "", "initial_priority"], [5, 4, 1, "", "is_link_preload"], [5, 4, 1, "", "is_same_site"], [5, 4, 1, "", "method"], [5, 4, 1, "", "mixed_content_type"], [5, 4, 1, "", "params"], [5, 4, 1, "", "post_data"], [5, 4, 1, "", "post_data_entries"], [5, 4, 1, "", "referrer_policy"], [5, 4, 1, "", "target"], [5, 4, 1, "", "trust_token_params"], [5, 4, 1, "", "url"], [5, 4, 1, "", "url_fragment"]], "selenium_driverless.scripts.network_interceptor.RequestDoneException": [[5, 4, 1, "", "request"]], "selenium_driverless.scripts.network_interceptor.RequestPattern": [[5, 2, 1, "", "AnyRequest"], [5, 2, 1, "", "AnyResponse"], [5, 1, 1, "", "new"]], "selenium_driverless.scripts.network_interceptor.RequestStages": [[5, 2, 1, "", "Request"], [5, 2, 1, "", "Response"]], "selenium_driverless.scripts.switch_to": [[1, 0, 1, "", "SwitchTo"]], "selenium_driverless.scripts.switch_to.SwitchTo": [[1, 4, 1, "", "alert"], [1, 1, 1, "", "default_content"], [1, 1, 1, "", "frame"], [1, 1, 1, "", "get_alert"], [1, 1, 1, "", "new_window"], [1, 1, 1, "", "target"], [1, 1, 1, "", "window"]], "selenium_driverless.types.base_target": [[6, 0, 1, "", "BaseTarget"]], "selenium_driverless.types.base_target.BaseTarget": [[6, 1, 1, "", "add_cdp_listener"], [6, 1, 1, "", "downloads_dir_for_context"], [6, 1, 1, "", "execute_cdp_cmd"], [6, 1, 1, "", "get_cdp_event_iter"], [6, 1, 1, "", "remove_cdp_listener"], [6, 4, 1, "", "socket"], [6, 1, 1, "", "wait_for_cdp"]], "selenium_driverless.types.by": [[0, 0, 1, "", "By"]], "selenium_driverless.types.by.By": [[0, 2, 1, "", "CLASS_NAME"], [0, 2, 1, "", "CSS"], [0, 2, 1, "", "CSS_SELECTOR"], [0, 2, 1, "", "ID"], [0, 2, 1, "", "NAME"], [0, 2, 1, "", "TAG_NAME"], [0, 2, 1, "", "XPATH"]], "selenium_driverless.types.context": [[3, 0, 1, "", "Context"]], "selenium_driverless.types.context.Context": [[3, 1, 1, "", "add_cdp_listener"], [3, 1, 1, "", "add_cookie"], [3, 1, 1, "", "back"], [3, 1, 1, "", "close"], [3, 4, 1, "", "current_pointer"], [3, 4, 1, "", "current_url"], [3, 4, 1, "", "current_window_handle"], [3, 1, 1, "", "delete_all_cookies"], [3, 1, 1, "", "delete_cookie"], [3, 1, 1, "", "delete_network_conditions"], [3, 4, 1, "", "downloads_dir"], [3, 1, 1, "", "eval_async"], [3, 1, 1, "", "execute_async_script"], [3, 1, 1, "", "execute_cdp_cmd"], [3, 1, 1, "", "execute_raw_script"], [3, 1, 1, "", "execute_script"], [3, 1, 1, "", "fetch"], [3, 1, 1, "", "find_element"], [3, 1, 1, "", "find_elements"], [3, 1, 1, "", "forward"], [3, 1, 1, "", "fullscreen_window"], [3, 1, 1, "", "get"], [3, 1, 1, "", "get_cdp_event_iter"], [3, 1, 1, "", "get_cookie"], [3, 1, 1, "", "get_cookies"], [3, 1, 1, "", "get_issue_message"], [3, 1, 1, "", "get_network_conditions"], [3, 1, 1, "", "get_screenshot_as_file"], [3, 1, 1, "", "get_screenshot_as_png"], [3, 1, 1, "", "get_sinks"], [3, 1, 1, "", "get_window_position"], [3, 1, 1, "", "get_window_rect"], [3, 1, 1, "", "get_window_size"], [3, 1, 1, "", "maximize_window"], [3, 1, 1, "", "minimize_window"], [3, 1, 1, "", "new_window"], [3, 4, 1, "", "page_source"], [3, 1, 1, "", "print_page"], [3, 1, 1, "", "quit"], [3, 1, 1, "", "refresh"], [3, 1, 1, "", "remove_cdp_listener"], [3, 1, 1, "", "save_screenshot"], [3, 1, 1, "", "save_snapshot"], [3, 1, 1, "", "search_elements"], [3, 1, 1, "", "send_keys"], [3, 1, 1, "", "set_download_behaviour"], [3, 1, 1, "", "set_network_conditions"], [3, 1, 1, "", "set_permissions"], [3, 1, 1, "", "set_sink_to_use"], [3, 1, 1, "", "set_window_position"], [3, 1, 1, "", "set_window_rect"], [3, 1, 1, "", "set_window_size"], [3, 1, 1, "", "snapshot"], [3, 1, 1, "", "start_desktop_mirroring"], [3, 1, 1, "", "start_tab_mirroring"], [3, 1, 1, "", "stop_casting"], [3, 4, 1, "", "switch_to"], [3, 4, 1, "", "title"], [3, 1, 1, "", "wait_download"], [3, 1, 1, "", "wait_for_cdp"], [3, 4, 1, "", "window_handles"], [3, 1, 1, "", "xhr"]], "selenium_driverless.types.options": [[2, 0, 1, "", "Options"]], "selenium_driverless.types.options.Options": [[2, 4, 1, "", "accept_insecure_certs"], [2, 1, 1, "", "add_argument"], [2, 1, 1, "", "add_arguments"], [2, 1, 1, "", "add_experimental_option"], [2, 1, 1, "", "add_extension"], [2, 4, 1, "", "arguments"], [2, 4, 1, "", "auto_clean_dirs"], [2, 4, 1, "", "binary_location"], [2, 4, 1, "", "debugger_address"], [2, 4, 1, "", "downloads_dir"], [2, 1, 1, "", "enable_mobile"], [2, 4, 1, "", "env"], [2, 4, 1, "", "headless"], [2, 1, 1, "", "ignore_local_proxy_environment_variables"], [2, 4, 1, "", "prefs"], [2, 4, 1, "", "single_proxy"], [2, 4, 1, "", "startup_url"], [2, 1, 1, "", "update_pref"], [2, 2, 1, "", "use_extension"], [2, 4, 1, "", "user_data_dir"]], "selenium_driverless.types.target": [[6, 0, 1, "", "Target"], [6, 0, 1, "", "TargetInfo"]], "selenium_driverless.types.target.Target": [[6, 1, 1, "", "activate"], [6, 1, 1, "", "add_cdp_listener"], [6, 1, 1, "", "add_cookie"], [6, 1, 1, "", "back"], [6, 1, 1, "", "close"], [6, 4, 1, "", "current_url"], [6, 1, 1, "", "delete_all_cookies"], [6, 1, 1, "", "delete_cookie"], [6, 1, 1, "", "delete_network_conditions"], [6, 1, 1, "", "eval_async"], [6, 1, 1, "", "execute_async_script"], [6, 1, 1, "", "execute_cdp_cmd"], [6, 1, 1, "", "execute_raw_script"], [6, 1, 1, "", "execute_script"], [6, 1, 1, "", "fetch"], [6, 1, 1, "", "find_element"], [6, 1, 1, "", "find_elements"], [6, 1, 1, "", "focus"], [6, 1, 1, "", "forward"], [6, 1, 1, "", "get"], [6, 1, 1, "", "get_cdp_event_iter"], [6, 1, 1, "", "get_cookie"], [6, 1, 1, "", "get_cookies"], [6, 1, 1, "", "get_history"], [6, 1, 1, "", "get_issue_message"], [6, 1, 1, "", "get_network_conditions"], [6, 1, 1, "", "get_screenshot_as_file"], [6, 1, 1, "", "get_screenshot_as_png"], [6, 1, 1, "", "get_sinks"], [6, 1, 1, "", "get_target_for_iframe"], [6, 1, 1, "", "get_targets_for_iframes"], [6, 4, 1, "", "page_source"], [6, 4, 1, "", "pointer"], [6, 1, 1, "", "print_page"], [6, 1, 1, "", "refresh"], [6, 1, 1, "", "remove_cdp_listener"], [6, 1, 1, "", "save_screenshot"], [6, 1, 1, "", "save_snapshot"], [6, 1, 1, "", "search_elements"], [6, 1, 1, "", "send_keys"], [6, 1, 1, "", "set_network_conditions"], [6, 1, 1, "", "set_sink_to_use"], [6, 1, 1, "", "set_source"], [6, 1, 1, "", "snapshot"], [6, 4, 1, "", "socket"], [6, 1, 1, "", "start_desktop_mirroring"], [6, 1, 1, "", "start_tab_mirroring"], [6, 1, 1, "", "stop_casting"], [6, 4, 1, "", "title"], [6, 1, 1, "", "unfocus"], [6, 1, 1, "", "wait_download"], [6, 1, 1, "", "wait_for_cdp"], [6, 1, 1, "", "xhr"]], "selenium_driverless.types.target.TargetInfo": [[6, 4, 1, "", "Target"], [6, 4, 1, "", "attached"], [6, 4, 1, "", "browser_context_id"], [6, 4, 1, "", "can_access_opener"], [6, 4, 1, "", "id"], [6, 4, 1, "", "opener_frame_id"], [6, 4, 1, "", "opener_id"], [6, 4, 1, "", "subtype"]], "selenium_driverless.types.webelement": [[7, 0, 1, "", "WebElement"]], "selenium_driverless.types.webelement.WebElement": [[7, 4, 1, "", "accessible_name"], [7, 4, 1, "", "aria_role"], [7, 4, 1, "", "backend_node_id"], [7, 4, 1, "", "box_model"], [7, 4, 1, "", "class_name"], [7, 1, 1, "", "clear"], [7, 1, 1, "", "click"], [7, 4, 1, "", "content_document"], [7, 4, 1, "", "context_id"], [7, 4, 1, "", "document_url"], [7, 4, 1, "", "dom_attributes"], [7, 1, 1, "", "eval_async"], [7, 1, 1, "", "execute_async_script"], [7, 1, 1, "", "execute_script"], [7, 1, 1, "", "find_element"], [7, 1, 1, "", "find_elements"], [7, 1, 1, "", "focus"], [7, 1, 1, "", "get_attribute"], [7, 1, 1, "", "get_dom_attribute"], [7, 1, 1, "", "get_listeners"], [7, 1, 1, "", "get_property"], [7, 1, 1, "", "highlight"], [7, 1, 1, "", "is_clickable"], [7, 1, 1, "", "is_displayed"], [7, 1, 1, "", "is_enabled"], [7, 1, 1, "", "is_selected"], [7, 4, 1, "", "location"], [7, 4, 1, "", "location_once_scrolled_into_view"], [7, 1, 1, "", "mid_location"], [7, 4, 1, "", "node_id"], [7, 4, 1, "", "obj_id"], [7, 4, 1, "", "parent"], [7, 4, 1, "", "rect"], [7, 1, 1, "", "remove"], [7, 1, 1, "", "screenshot"], [7, 4, 1, "", "screenshot_as_base64"], [7, 4, 1, "", "screenshot_as_png"], [7, 1, 1, "", "scroll_to"], [7, 1, 1, "", "send_keys"], [7, 1, 1, "", "set_dom_attribute"], [7, 1, 1, "", "set_file"], [7, 1, 1, "", "set_files"], [7, 1, 1, "", "set_source"], [7, 4, 1, "", "shadow_root"], [7, 4, 1, "", "shadow_roots"], [7, 4, 1, "", "size"], [7, 4, 1, "", "source"], [7, 1, 1, "", "submit"], [7, 4, 1, "", "tag_name"], [7, 4, 1, "", "text"], [7, 4, 1, "", "value"], [7, 1, 1, "", "value_of_css_property"], [7, 1, 1, "", "write"]], "selenium_driverless.webdriver": [[1, 0, 1, "", "Chrome"]], "selenium_driverless.webdriver.Chrome": [[1, 1, 1, "", "add_cdp_listener"], [1, 1, 1, "", "add_cookie"], [1, 1, 1, "", "back"], [1, 4, 1, "", "base_context"], [1, 4, 1, "", "base_target"], [1, 1, 1, "", "clear_auth"], [1, 1, 1, "", "clear_proxy"], [1, 1, 1, "", "close"], [1, 4, 1, "", "contexts"], [1, 4, 1, "", "current_context"], [1, 4, 1, "", "current_pointer"], [1, 4, 1, "", "current_target"], [1, 4, 1, "", "current_target_info"], [1, 4, 1, "", "current_url"], [1, 4, 1, "", "current_window_handle"], [1, 4, 1, "", "current_window_id"], [1, 1, 1, "", "delete_all_cookies"], [1, 1, 1, "", "delete_cookie"], [1, 1, 1, "", "delete_network_conditions"], [1, 4, 1, "", "downloads_dir"], [1, 1, 1, "", "ensure_extensions_incognito_allowed"], [1, 1, 1, "", "eval_async"], [1, 1, 1, "", "execute_async_script"], [1, 1, 1, "", "execute_cdp_cmd"], [1, 1, 1, "", "execute_raw_script"], [1, 1, 1, "", "execute_script"], [1, 1, 1, "", "fetch"], [1, 1, 1, "", "find_element"], [1, 1, 1, "", "find_elements"], [1, 1, 1, "", "focus"], [1, 1, 1, "", "forward"], [1, 4, 1, "", "frame_tree"], [1, 1, 1, "", "fullscreen_window"], [1, 1, 1, "", "get"], [1, 1, 1, "", "get_cdp_event_iter"], [1, 1, 1, "", "get_cookie"], [1, 1, 1, "", "get_cookies"], [1, 1, 1, "", "get_issue_message"], [1, 1, 1, "", "get_network_conditions"], [1, 1, 1, "", "get_screenshot_as_file"], [1, 1, 1, "", "get_screenshot_as_png"], [1, 1, 1, "", "get_sinks"], [1, 1, 1, "", "get_target"], [1, 1, 1, "", "get_target_for_iframe"], [1, 1, 1, "", "get_targets"], [1, 1, 1, "", "get_targets_for_iframes"], [1, 1, 1, "", "get_window_position"], [1, 1, 1, "", "get_window_rect"], [1, 1, 1, "", "get_window_size"], [1, 1, 1, "", "maximize_window"], [1, 1, 1, "", "minimize_window"], [1, 4, 1, "", "mv3_extension"], [1, 1, 1, "", "new_context"], [1, 1, 1, "", "new_window"], [1, 1, 1, "", "normalize_window"], [1, 4, 1, "", "page_source"], [1, 1, 1, "", "print_page"], [1, 1, 1, "", "quit"], [1, 1, 1, "", "refresh"], [1, 1, 1, "", "remove_cdp_listener"], [1, 1, 1, "", "save_screenshot"], [1, 1, 1, "", "save_snapshot"], [1, 1, 1, "", "search_elements"], [1, 1, 1, "", "send_keys"], [1, 1, 1, "", "set_auth"], [1, 1, 1, "", "set_download_behaviour"], [1, 1, 1, "", "set_network_conditions"], [1, 1, 1, "", "set_permissions"], [1, 1, 1, "", "set_proxy"], [1, 1, 1, "", "set_single_proxy"], [1, 1, 1, "", "set_sink_to_use"], [1, 1, 1, "", "set_window_position"], [1, 1, 1, "", "set_window_rect"], [1, 1, 1, "", "set_window_size"], [1, 1, 1, "", "set_window_state"], [1, 1, 1, "", "sleep"], [1, 1, 1, "", "snapshot"], [1, 1, 1, "", "start_desktop_mirroring"], [1, 1, 1, "", "start_tab_mirroring"], [1, 1, 1, "", "stop_casting"], [1, 4, 1, "", "switch_to"], [1, 4, 1, "", "targets"], [1, 4, 1, "", "title"], [1, 1, 1, "", "wait_download"], [1, 1, 1, "", "wait_for_cdp"], [1, 4, 1, "", "window_handles"], [1, 1, 1, "", "xhr"]]}, "objtypes": {"0": "py:class", "1": "py:method", "2": "py:attribute", "3": "py:function", "4": "py:property"}, "objnames": {"0": ["py", "class", "Python class"], "1": ["py", "method", "Python method"], "2": ["py", "attribute", "Python attribute"], "3": ["py", "function", "Python function"], "4": ["py", "property", "Python property"]}, "titleterms": {"By": 0, "element": [0, 4], "locat": 0, "webdriv": 1, "chrome": 1, "chromeopt": 2, "context": 3, "input": 4, "request": 5, "intercept": 5, "exampl": 5, "script": 5, "api": [5, 8], "target": 6, "webel": 7, "document": 8, "driverless": 8, "instal": 8, "usag": 8, "sourc": 8, "select": 4}, "envversion": {"sphinx.domains.c": 3, "sphinx.domains.changeset": 1, "sphinx.domains.citation": 1, "sphinx.domains.cpp": 9, "sphinx.domains.index": 1, "sphinx.domains.javascript": 3, "sphinx.domains.math": 2, "sphinx.domains.python": 4, "sphinx.domains.rst": 2, "sphinx.domains.std": 2, "sphinx.ext.viewcode": 1, "sphinx": 58}, "alltitles": {"By Element Locator": [[0, "by-element-locator"]], "ChromeOptions": [[2, "chromeoptions"]], "Request-Interception": [[5, "request-interception"]], "Example Script": [[5, "example-script"]], "API": [[5, "api"], [8, "api"]], "WebElement": [[7, "webelement"]], "Documentation of Driverless": [[8, "documentation-of-driverless"]], "Installation": [[8, "installation"]], "Usage": [[8, "usage"]], "Source": [[8, "source"]], "webdriver.Chrome": [[1, "webdriver-chrome"]], "Context": [[3, "context"]], "Target": [[6, "target"]], "Input": [[4, "input"]], "Select Element": [[4, "select-element"]]}, "indexentries": {"alt (selenium_driverless.input.pointer.modifiers attribute)": [[4, "selenium_driverless.input.pointer.Modifiers.ALT"]], "back (selenium_driverless.input.pointer.buttons attribute)": [[4, "selenium_driverless.input.pointer.Buttons.BACK"]], "back (selenium_driverless.input.pointer.mousebutton attribute)": [[4, "selenium_driverless.input.pointer.MouseButton.BACK"]], "basepointer (class in selenium_driverless.input.pointer)": [[4, "selenium_driverless.input.pointer.BasePointer"]], "buttons (class in selenium_driverless.input.pointer)": [[4, "selenium_driverless.input.pointer.Buttons"]], "command (selenium_driverless.input.pointer.modifiers attribute)": [[4, "selenium_driverless.input.pointer.Modifiers.COMMAND"]], "ctrl (selenium_driverless.input.pointer.modifiers attribute)": [[4, "selenium_driverless.input.pointer.Modifiers.CTRL"]], "default (selenium_driverless.input.pointer.buttons attribute)": [[4, "selenium_driverless.input.pointer.Buttons.DEFAULT"]], "eventtype (class in selenium_driverless.input.pointer)": [[4, "selenium_driverless.input.pointer.EventType"]], "forward (selenium_driverless.input.pointer.buttons attribute)": [[4, "selenium_driverless.input.pointer.Buttons.FORWARD"]], "forward (selenium_driverless.input.pointer.mousebutton attribute)": [[4, "selenium_driverless.input.pointer.MouseButton.FORWARD"]], "left (selenium_driverless.input.pointer.buttons attribute)": [[4, "selenium_driverless.input.pointer.Buttons.LEFT"]], "left (selenium_driverless.input.pointer.mousebutton attribute)": [[4, "selenium_driverless.input.pointer.MouseButton.LEFT"]], "middle (selenium_driverless.input.pointer.buttons attribute)": [[4, "selenium_driverless.input.pointer.Buttons.MIDDLE"]], "middle (selenium_driverless.input.pointer.mousebutton attribute)": [[4, "selenium_driverless.input.pointer.MouseButton.MIDDLE"]], "mouse (selenium_driverless.input.pointer.pointertype attribute)": [[4, "selenium_driverless.input.pointer.PointerType.MOUSE"]], "move (selenium_driverless.input.pointer.eventtype attribute)": [[4, "selenium_driverless.input.pointer.EventType.MOVE"]], "modifiers (class in selenium_driverless.input.pointer)": [[4, "selenium_driverless.input.pointer.Modifiers"]], "mousebutton (class in selenium_driverless.input.pointer)": [[4, "selenium_driverless.input.pointer.MouseButton"]], "none (selenium_driverless.input.pointer.buttons attribute)": [[4, "selenium_driverless.input.pointer.Buttons.NONE"]], "none (selenium_driverless.input.pointer.modifiers attribute)": [[4, "selenium_driverless.input.pointer.Modifiers.NONE"]], "none (selenium_driverless.input.pointer.mousebutton attribute)": [[4, "selenium_driverless.input.pointer.MouseButton.NONE"]], "pen (selenium_driverless.input.pointer.pointertype attribute)": [[4, "selenium_driverless.input.pointer.PointerType.PEN"]], "press (selenium_driverless.input.pointer.eventtype attribute)": [[4, "selenium_driverless.input.pointer.EventType.PRESS"]], "pointer (class in selenium_driverless.input.pointer)": [[4, "selenium_driverless.input.pointer.Pointer"]], "pointerevent (class in selenium_driverless.input.pointer)": [[4, "selenium_driverless.input.pointer.PointerEvent"]], "pointertype (class in selenium_driverless.input.pointer)": [[4, "selenium_driverless.input.pointer.PointerType"]], "release (selenium_driverless.input.pointer.eventtype attribute)": [[4, "selenium_driverless.input.pointer.EventType.RELEASE"]], "right (selenium_driverless.input.pointer.buttons attribute)": [[4, "selenium_driverless.input.pointer.Buttons.RIGHT"]], "right (selenium_driverless.input.pointer.mousebutton attribute)": [[4, "selenium_driverless.input.pointer.MouseButton.RIGHT"]], "shift (selenium_driverless.input.pointer.modifiers attribute)": [[4, "selenium_driverless.input.pointer.Modifiers.SHIFT"]], "wheel (selenium_driverless.input.pointer.eventtype attribute)": [[4, "selenium_driverless.input.pointer.EventType.WHEEL"]], "click() (selenium_driverless.input.pointer.basepointer method)": [[4, "selenium_driverless.input.pointer.BasePointer.click"]], "click() (selenium_driverless.input.pointer.pointer method)": [[4, "selenium_driverless.input.pointer.Pointer.click"]], "dispatch() (selenium_driverless.input.pointer.basepointer method)": [[4, "selenium_driverless.input.pointer.BasePointer.dispatch"]], "double_click() (selenium_driverless.input.pointer.basepointer method)": [[4, "selenium_driverless.input.pointer.BasePointer.double_click"]], "down() (selenium_driverless.input.pointer.basepointer method)": [[4, "selenium_driverless.input.pointer.BasePointer.down"]], "down() (selenium_driverless.input.pointer.pointer method)": [[4, "selenium_driverless.input.pointer.Pointer.down"]], "move_path() (selenium_driverless.input.pointer.basepointer method)": [[4, "selenium_driverless.input.pointer.BasePointer.move_path"]], "move_to() (selenium_driverless.input.pointer.basepointer method)": [[4, "selenium_driverless.input.pointer.BasePointer.move_to"]], "move_to() (selenium_driverless.input.pointer.pointer method)": [[4, "selenium_driverless.input.pointer.Pointer.move_to"]], "select() (in module selenium_driverless.input.utils)": [[4, "selenium_driverless.input.utils.select"]], "to_json() (selenium_driverless.input.pointer.pointerevent method)": [[4, "selenium_driverless.input.pointer.PointerEvent.to_json"]], "up() (selenium_driverless.input.pointer.basepointer method)": [[4, "selenium_driverless.input.pointer.BasePointer.up"]], "up() (selenium_driverless.input.pointer.pointer method)": [[4, "selenium_driverless.input.pointer.Pointer.up"]]}})
