+++ tests/html/test_select.py
@@ -3,11 +3,9 @@ import asyncio
 from selenium_driverless.types.target import Target
 from selenium_driverless.webdriver import Chrome
 from selenium_driverless.types.by import By
-from selenium_driverless.types.webelement import WebElement
-from selenium_driverless.types import JSEvalException
+from selenium_driverless.input.utils import select
 
 from cdp_patches.input import AsyncInput
-from cdp_patches.input import KeyboardCodes
 
 # puppeteer:
 # https://github.com/puppeteer/puppeteer/blob/409d244aed480fbb5254f852afb16bd101692f9a/packages/puppeteer-core/src/api/ElementHandle.ts#L919-L961
@@ -102,61 +100,6 @@ async def down(tab: Target):
 async def add_elem(driver: Chrome):
     await driver.current_target.set_source(select_html)
     await driver.execute_script(track_js, unique_context=False)
-
-
-async def select(elem: WebElement, value: str = None, text: str = None, async_input: AsyncInput = None,
-                 timeouts: float = 0.01, up: bool = False) -> None:
-    """
-    selects an option of a select element
-    """
-    use_js = async_input is None
-    if use_js:
-        await elem.click()
-    else:
-        x, y = await elem.mid_location()
-        await async_input.click("left", x, y)
-    if value is None and text is None:
-        raise ValueError("value or text need to be specified")
-    try:
-        n, direction = await elem.execute_script("""
-            let [value, text, use_js] = arguments
-            var idx = Array.from(obj.options).findIndex(option => option.value === value || option.text === text)
-            var currIdx = obj.selectedIndex
-            if (idx === -1){throw ReferenceError("option not found")}
-            if(typeof value == 'undefined'){
-                value = obj.options[idx].value
-            }
-            if(use_js && obj.options[currIdx].value !== value){
-                obj.value = value
-                const evt = new Event("change")
-                evt.initEvent("change", true, true)
-                obj.dispatchEvent(evt)
-                return [0, 1]
-            }else{
-                const n = Math.abs(idx - currIdx);
-                const direction = idx < currIdx ? 1 : -1
-                return [n, direction]
-            }
-            
-        """, value, text, use_js)
-    except JSEvalException as e:
-        if e.class_name == "ReferenceError" and e.description[:33] == 'ReferenceError: option not found\n':
-            raise ValueError(f"option not found based on value:{value}, text:{text} for {elem}")
-        raise e
-    if use_js is False:
-        if up:
-            direction *= -1
-        if direction == 1:
-            code = KeyboardCodes.UP_ARROW
-        else:
-            code = KeyboardCodes.DOWN_ARROW
-        for _ in range(n):
-            await asyncio.sleep(timeouts)
-            async_input.base.send_keystrokes(code)
-        await asyncio.sleep(timeouts)
-        x, y = await elem.mid_location()
-        await async_input.click("left", x, y)
-        await asyncio.sleep(timeouts)
 
 
 async def select_test(driver, subtests, headfull=False):
